@@ -1,44 +1,30 @@
 // Code generated by Wire. DO NOT EDIT.
 
-//go:generate wire
-//+build !wireinject
+//go:generate go run github.com/google/wire/cmd/wire
+//go:build !wireinject
+// +build !wireinject
 
 package server
 
 import (
-	"github.com/aquasecurity/fanal/applier"
-	"github.com/aquasecurity/fanal/cache"
 	"github.com/aquasecurity/trivy-db/pkg/db"
-	db2 "github.com/aquasecurity/trivy/pkg/db"
-	"github.com/aquasecurity/trivy/pkg/detector/ospkg"
-	"github.com/aquasecurity/trivy/pkg/github"
-	"github.com/aquasecurity/trivy/pkg/indicator"
-	"github.com/aquasecurity/trivy/pkg/result"
+	"github.com/aquasecurity/trivy/pkg/cache"
+	"github.com/aquasecurity/trivy/pkg/fanal/applier"
+	"github.com/aquasecurity/trivy/pkg/scanner/langpkg"
 	"github.com/aquasecurity/trivy/pkg/scanner/local"
-	"github.com/spf13/afero"
-	"k8s.io/utils/clock"
+	"github.com/aquasecurity/trivy/pkg/scanner/ospkg"
+	"github.com/aquasecurity/trivy/pkg/vulnerability"
 )
 
 // Injectors from inject.go:
 
 func initializeScanServer(localArtifactCache cache.LocalArtifactCache) *ScanServer {
 	applierApplier := applier.NewApplier(localArtifactCache)
-	detector := ospkg.Detector{}
-	scanner := local.NewScanner(applierApplier, detector)
+	scanner := ospkg.NewScanner()
+	langpkgScanner := langpkg.NewScanner()
 	config := db.Config{}
-	client := result.NewClient(config)
-	scanServer := NewScanServer(scanner, client)
+	client := vulnerability.NewClient(config)
+	localScanner := local.NewScanner(applierApplier, scanner, langpkgScanner, client)
+	scanServer := NewScanServer(localScanner)
 	return scanServer
-}
-
-func initializeDBWorker(cacheDir string, quiet bool) dbWorker {
-	config := db.Config{}
-	client := github.NewClient()
-	progressBar := indicator.NewProgressBar(quiet)
-	realClock := clock.RealClock{}
-	fs := afero.NewOsFs()
-	metadata := db2.NewMetadata(fs, cacheDir)
-	dbClient := db2.NewClient(config, client, progressBar, realClock, metadata)
-	serverDbWorker := newDBWorker(dbClient)
-	return serverDbWorker
 }
