@@ -2,16 +2,26 @@ package cache
 
 import (
 	"context"
+	"crypto/tls"
 	"net/http"
 
+	"github.com/twitchtv/twirp"
 	"golang.org/x/xerrors"
 
-	"github.com/aquasecurity/fanal/cache"
-	"github.com/aquasecurity/fanal/types"
+	"github.com/aquasecurity/trivy/pkg/fanal/types"
 	"github.com/aquasecurity/trivy/pkg/rpc"
 	"github.com/aquasecurity/trivy/pkg/rpc/client"
 	rpcCache "github.com/aquasecurity/trivy/rpc/cache"
 )
+
+var _ ArtifactCache = (*RemoteCache)(nil)
+
+type RemoteOptions struct {
+	ServerAddr    string
+	CustomHeaders http.Header
+	Insecure      bool
+	PathPrefix    string
+}
 
 // RemoteCache implements remote cache
 type RemoteCache struct {
@@ -19,19 +29,37 @@ type RemoteCache struct {
 	client rpcCache.Cache
 }
 
-// RemoteURL to hold remote host
-type RemoteURL string
-
 // NewRemoteCache is the factory method for RemoteCache
-func NewRemoteCache(url RemoteURL, customHeaders http.Header) cache.ArtifactCache {
-	ctx := client.WithCustomHeaders(context.Background(), customHeaders)
-	c := rpcCache.NewCacheProtobufClient(string(url), &http.Client{})
-	return &RemoteCache{ctx: ctx, client: c}
+func NewRemoteCache(opts RemoteOptions) *RemoteCache {
+	ctx := client.WithCustomHeaders(context.Background(), opts.CustomHeaders)
+
+	httpClient := &http.Client{
+		Transport: &http.Transport{
+			Proxy: http.ProxyFromEnvironment,
+			TLSClientConfig: &tls.Config{
+				InsecureSkipVerify: opts.Insecure,
+			},
+		},
+	}
+
+	var twirpOpts []twirp.ClientOption
+	if opts.PathPrefix != "" {
+		twirpOpts = append(twirpOpts, twirp.WithClientPathPrefix(opts.PathPrefix))
+	}
+	c := rpcCache.NewCacheProtobufClient(opts.ServerAddr, httpClient, twirpOpts...)
+	return &RemoteCache{
+		ctx:    ctx,
+		client: c,
+	}
 }
 
 // PutArtifact sends artifact to remote client
 func (c RemoteCache) PutArtifact(imageID string, artifactInfo types.ArtifactInfo) error {
-	_, err := c.client.PutArtifact(c.ctx, rpc.ConvertToRPCArtifactInfo(imageID, artifactInfo))
+	err := rpc.Retry(func() error {
+		var err error
+		_, err = c.client.PutArtifact(c.ctx, rpc.ConvertToRPCArtifactInfo(imageID, artifactInfo))
+		return err
+	})
 	if err != nil {
 		return xerrors.Errorf("unable to store cache on the server: %w", err)
 	}
@@ -40,7 +68,11 @@ func (c RemoteCache) PutArtifact(imageID string, artifactInfo types.ArtifactInfo
 
 // PutBlob sends blobInfo to remote client
 func (c RemoteCache) PutBlob(diffID string, blobInfo types.BlobInfo) error {
-	_, err := c.client.PutBlob(c.ctx, rpc.ConvertToRPCBlobInfo(diffID, blobInfo))
+	err := rpc.Retry(func() error {
+		var err error
+		_, err = c.client.PutBlob(c.ctx, rpc.ConvertToRPCPutBlobRequest(diffID, blobInfo))
+		return err
+	})
 	if err != nil {
 		return xerrors.Errorf("unable to store cache on the server: %w", err)
 	}
@@ -49,9 +81,27 @@ func (c RemoteCache) PutBlob(diffID string, blobInfo types.BlobInfo) error {
 
 // MissingBlobs fetches missing blobs from RemoteCache
 func (c RemoteCache) MissingBlobs(imageID string, layerIDs []string) (bool, []string, error) {
-	layers, err := c.client.MissingBlobs(c.ctx, rpc.ConvertToMissingBlobsRequest(imageID, layerIDs))
+	var layers *rpcCache.MissingBlobsResponse
+	err := rpc.Retry(func() error {
+		var err error
+		layers, err = c.client.MissingBlobs(c.ctx, rpc.ConvertToMissingBlobsRequest(imageID, layerIDs))
+		return err
+	})
 	if err != nil {
 		return false, nil, xerrors.Errorf("unable to fetch missing layers: %w", err)
 	}
 	return layers.MissingArtifact, layers.MissingBlobIds, nil
+}
+
+// DeleteBlobs removes blobs by IDs from RemoteCache
+func (c RemoteCache) DeleteBlobs(blobIDs []string) error {
+	err := rpc.Retry(func() error {
+		var err error
+		_, err = c.client.DeleteBlobs(c.ctx, rpc.ConvertToDeleteBlobsRequest(blobIDs))
+		return err
+	})
+	if err != nil {
+		return xerrors.Errorf("unable to delete blobs on the server: %w", err)
+	}
+	return nil
 }
