@@ -1,57 +1,55 @@
 package server
 
 import (
-	"github.com/urfave/cli/v2"
+	"context"
+
 	"golang.org/x/xerrors"
 
-	"github.com/aquasecurity/trivy-db/pkg/db"
+	"github.com/aquasecurity/trivy/pkg/cache"
 	"github.com/aquasecurity/trivy/pkg/commands/operation"
+	"github.com/aquasecurity/trivy/pkg/db"
+	"github.com/aquasecurity/trivy/pkg/flag"
 	"github.com/aquasecurity/trivy/pkg/log"
+	"github.com/aquasecurity/trivy/pkg/module"
 	rpcServer "github.com/aquasecurity/trivy/pkg/rpc/server"
-	"github.com/aquasecurity/trivy/pkg/utils"
 )
 
 // Run runs the scan
-func Run(ctx *cli.Context) error {
-	return run(NewConfig(ctx))
-}
-
-func run(c Config) (err error) {
-	if err = log.InitLogger(c.Debug, c.Quiet); err != nil {
-		return xerrors.Errorf("failed to initialize a logger: %w", err)
-	}
-
-	// initialize config
-	if err = c.Init(); err != nil {
-		return xerrors.Errorf("failed to initialize options: %w", err)
-	}
+func Run(ctx context.Context, opts flag.Options) (err error) {
+	log.InitLogger(opts.Debug, opts.Quiet)
 
 	// configure cache dir
-	utils.SetCacheDir(c.CacheDir)
-	cache, err := operation.NewCache(c.CacheBackend)
+	cacheClient, cleanup, err := cache.New(opts.CacheOpts())
 	if err != nil {
 		return xerrors.Errorf("server cache error: %w", err)
 	}
-	defer cache.Close()
-	log.Logger.Debugf("cache dir:  %s", utils.CacheDir())
-
-	if c.Reset {
-		return cache.ClearDB()
-	}
+	defer cleanup()
 
 	// download the database file
-	if err = operation.DownloadDB(c.AppVersion, c.CacheDir, true, false, c.SkipDBUpdate); err != nil {
+	if err = operation.DownloadDB(ctx, opts.AppVersion, opts.CacheDir, opts.DBRepositories,
+		true, opts.SkipDBUpdate, opts.RegistryOpts()); err != nil {
 		return err
 	}
 
-	if c.DownloadDBOnly {
+	if opts.DownloadDBOnly {
 		return nil
 	}
 
-	if err = db.Init(c.CacheDir); err != nil {
+	if err = db.Init(db.Dir(opts.CacheDir)); err != nil {
 		return xerrors.Errorf("error in vulnerability DB initialize: %w", err)
 	}
 
-	server := rpcServer.NewServer(c.AppVersion, c.Listen, c.CacheDir, c.Token, c.TokenHeader)
-	return server.ListenAndServe(cache)
+	// Initialize WASM modules
+	m, err := module.NewManager(ctx, module.Options{
+		Dir:            opts.ModuleDir,
+		EnabledModules: opts.EnabledModules,
+	})
+	if err != nil {
+		return xerrors.Errorf("WASM module error: %w", err)
+	}
+	m.Register()
+
+	server := rpcServer.NewServer(opts.AppVersion, opts.Listen, opts.CacheDir, opts.Token, opts.TokenHeader,
+		opts.PathPrefix, opts.DBRepositories, opts.RegistryOpts())
+	return server.ListenAndServe(ctx, cacheClient, opts.SkipDBUpdate)
 }
