@@ -1,268 +1,225 @@
 package db
 
 import (
-	"compress/gzip"
 	"context"
-	"encoding/json"
-	"io"
+	"fmt"
 	"os"
 	"path/filepath"
 	"time"
 
-	"github.com/google/wire"
-	"github.com/spf13/afero"
+	"github.com/google/go-containerregistry/pkg/name"
+	"github.com/samber/lo"
 	"golang.org/x/xerrors"
-	"k8s.io/utils/clock"
 
 	"github.com/aquasecurity/trivy-db/pkg/db"
-	"github.com/aquasecurity/trivy/pkg/github"
-	"github.com/aquasecurity/trivy/pkg/indicator"
+	"github.com/aquasecurity/trivy-db/pkg/metadata"
+	"github.com/aquasecurity/trivy/pkg/clock"
+	"github.com/aquasecurity/trivy/pkg/fanal/types"
 	"github.com/aquasecurity/trivy/pkg/log"
+	"github.com/aquasecurity/trivy/pkg/oci"
 )
 
 const (
-	fullDB  = "trivy.db.gz"
-	lightDB = "trivy-light.db.gz"
-
-	metadataFile = "metadata.json"
-
-	gb = 1024 * 1024 * 1024
+	SchemaVersion = db.SchemaVersion
+	dbMediaType   = "application/vnd.aquasec.trivy.db.layer.v1.tar+gzip"
 )
 
-// SuperSet binds the dependencies
-var SuperSet = wire.NewSet(
-	// indicator.ProgressBar
-	indicator.NewProgressBar,
+var (
+	// GitHub Container Registry
+	DefaultGHCRRepository = fmt.Sprintf("%s:%d", "ghcr.io/aquasecurity/trivy-db", db.SchemaVersion)
+	defaultGHCRRepository = lo.Must(name.NewTag(DefaultGHCRRepository))
 
-	// clock.Clock
-	wire.Struct(new(clock.RealClock)),
-	wire.Bind(new(clock.Clock), new(clock.RealClock)),
+	// GCR mirror
+	DefaultGCRRepository = fmt.Sprintf("%s:%d", "mirror.gcr.io/aquasec/trivy-db", db.SchemaVersion)
+	defaultGCRRepository = lo.Must(name.NewTag(DefaultGCRRepository))
 
-	// db.Config
-	wire.Struct(new(db.Config)),
-	wire.Bind(new(dbOperation), new(db.Config)),
-
-	// github.Client
-	github.NewClient,
-	wire.Bind(new(github.Operation), new(github.Client)),
-
-	// Metadata
-	afero.NewOsFs,
-	NewMetadata,
-
-	// db.Client
-	NewClient,
-	wire.Bind(new(Operation), new(Client)),
+	Init  = db.Init
+	Close = db.Close
+	Path  = db.Path
 )
 
-// Operation defines the DB operations
-type Operation interface {
-	NeedsUpdate(cliVersion string, light, skip bool) (need bool, err error)
-	Download(ctx context.Context, cacheDir string, light bool) (err error)
-	UpdateMetadata(cacheDir string) (err error)
+type options struct {
+	artifact       *oci.Artifact
+	dbRepositories []name.Reference
 }
 
-type dbOperation interface {
-	GetMetadata() (metadata db.Metadata, err error)
-	StoreMetadata(metadata db.Metadata, dir string) (err error)
+// Option is a functional option
+type Option func(*options)
+
+// WithOCIArtifact takes an OCI artifact
+func WithOCIArtifact(art *oci.Artifact) Option {
+	return func(opts *options) {
+		opts.artifact = art
+	}
+}
+
+// WithDBRepository takes a dbRepository
+func WithDBRepository(dbRepository []name.Reference) Option {
+	return func(opts *options) {
+		opts.dbRepositories = dbRepository
+	}
 }
 
 // Client implements DB operations
 type Client struct {
-	dbc          dbOperation
-	githubClient github.Operation
-	pb           indicator.ProgressBar
-	clock        clock.Clock
-	metadata     Metadata
+	*options
+
+	dbDir    string
+	metadata metadata.Client
+	quiet    bool
+}
+
+func Dir(cacheDir string) string {
+	return filepath.Join(cacheDir, "db")
 }
 
 // NewClient is the factory method for DB client
-func NewClient(dbc dbOperation, githubClient github.Operation, pb indicator.ProgressBar, clock clock.Clock, metadata Metadata) Client {
-	return Client{
-		dbc:          dbc,
-		githubClient: githubClient,
-		pb:           pb,
-		clock:        clock,
-		metadata:     metadata,
+func NewClient(dbDir string, quiet bool, opts ...Option) *Client {
+	o := &options{
+		dbRepositories: []name.Reference{
+			defaultGCRRepository,
+			defaultGHCRRepository,
+		},
+	}
+
+	for _, opt := range opts {
+		opt(o)
+	}
+
+	return &Client{
+		options:  o,
+		dbDir:    dbDir,
+		metadata: metadata.NewClient(dbDir),
+		quiet:    quiet,
 	}
 }
 
 // NeedsUpdate check is DB needs update
-func (c Client) NeedsUpdate(cliVersion string, light, skip bool) (bool, error) {
-	dbType := db.TypeFull
-	if light {
-		dbType = db.TypeLight
-	}
-
-	metadata, err := c.metadata.Get()
+func (c *Client) NeedsUpdate(ctx context.Context, cliVersion string, skip bool) (bool, error) {
+	meta, err := c.metadata.Get()
 	if err != nil {
-		log.Logger.Debugf("There is no valid metadata file: %s", err)
+		log.DebugContext(ctx, "There is no valid metadata file", log.Err(err))
 		if skip {
-			log.Logger.Error("The first run cannot skip downloading DB")
+			log.ErrorContext(ctx, "The first run cannot skip downloading DB")
 			return false, xerrors.New("--skip-update cannot be specified on the first run")
 		}
-		metadata = db.Metadata{} // suppress a warning
+		meta = metadata.Metadata{Version: db.SchemaVersion}
 	}
 
-	if db.SchemaVersion < metadata.Version {
-		log.Logger.Errorf("Trivy version (%s) is old. Update to the latest version.", cliVersion)
+	if db.SchemaVersion < meta.Version {
+		log.ErrorContext(ctx, "Trivy version is old. Update to the latest version.", log.String("version", cliVersion))
 		return false, xerrors.Errorf("the version of DB schema doesn't match. Local DB: %d, Expected: %d",
-			metadata.Version, db.SchemaVersion)
+			meta.Version, db.SchemaVersion)
 	}
 
 	if skip {
-		if err = c.validate(dbType, metadata); err != nil {
-			return false, err
+		log.DebugContext(ctx, "Skipping DB update...")
+		if err = c.validate(meta); err != nil {
+			return false, xerrors.Errorf("validate error: %w", err)
 		}
 		return false, nil
 	}
 
-	if db.SchemaVersion != metadata.Version || metadata.Type != dbType {
+	if db.SchemaVersion != meta.Version {
+		log.DebugContext(ctx, "The local DB schema version does not match with supported version schema.",
+			log.Int("local_version", meta.Version), log.Int("supported_version", db.SchemaVersion))
 		return true, nil
 	}
 
-	return !c.isNewDB(metadata), nil
+	return !c.isNewDB(ctx, meta), nil
 }
 
-func (c Client) validate(dbType db.Type, metadata db.Metadata) error {
-	if db.SchemaVersion != metadata.Version {
-		log.Logger.Error("The local DB is old and needs to be updated")
-		return xerrors.New("--skip-update cannot be specified with the old DB")
-	} else if metadata.Type != dbType {
-		if dbType == db.TypeFull {
-			log.Logger.Error("The local DB is a lightweight DB. You have to download a full DB")
-		} else {
-			log.Logger.Error("The local DB is a full DB. You have to download a lightweight DB")
-		}
-		return xerrors.New("--skip-update cannot be specified with the different schema DB")
+func (c *Client) validate(meta metadata.Metadata) error {
+	if db.SchemaVersion != meta.Version {
+		log.Error("The local DB has an old schema version which is not supported by the current version of Trivy CLI. DB needs to be updated.")
+		return xerrors.Errorf("--skip-update cannot be specified with the old DB schema. Local DB: %d, Expected: %d",
+			meta.Version, db.SchemaVersion)
 	}
 	return nil
 }
 
-func (c Client) isNewDB(metadata db.Metadata) bool {
-	if c.clock.Now().Before(metadata.NextUpdate) {
-		log.Logger.Debug("DB update was skipped because DB is the latest")
+func (c *Client) isNewDB(ctx context.Context, meta metadata.Metadata) bool {
+	now := clock.Now(ctx)
+	if now.Before(meta.NextUpdate) {
+		log.Debug("DB update was skipped because the local DB is the latest")
 		return true
 	}
 
-	if c.clock.Now().Before(metadata.DownloadedAt.Add(time.Hour)) {
-		log.Logger.Debug("DB update was skipped because DB was downloaded during the last hour")
+	if now.Before(meta.DownloadedAt.Add(time.Hour)) {
+		log.Debug("DB update was skipped because the local DB was downloaded during the last hour")
 		return true
 	}
 	return false
 }
 
 // Download downloads the DB file
-func (c Client) Download(ctx context.Context, cacheDir string, light bool) error {
-	// Remove the metadata file before downloading DB
+func (c *Client) Download(ctx context.Context, dst string, opt types.RegistryOptions) error {
+	// Remove the metadata file under the cache directory before downloading DB
 	if err := c.metadata.Delete(); err != nil {
-		log.Logger.Debug("no metadata file")
+		log.DebugContext(ctx, "No metadata file")
 	}
 
-	dbFile := fullDB
-	if light {
-		dbFile = lightDB
+	if err := c.downloadDB(ctx, opt, dst); err != nil {
+		return xerrors.Errorf("OCI artifact error: %w", err)
 	}
 
-	rc, size, err := c.githubClient.DownloadDB(ctx, dbFile)
-	if err != nil {
-		return xerrors.Errorf("failed to download vulnerability DB: %w", err)
+	if err := c.updateDownloadedAt(ctx, dst); err != nil {
+		return xerrors.Errorf("failed to update downloaded_at: %w", err)
 	}
-	defer rc.Close()
-
-	bar := c.pb.Start(int64(size))
-	barReader := bar.NewProxyReader(rc)
-	defer bar.Finish()
-	gr, err := gzip.NewReader(barReader)
-	if err != nil {
-		return xerrors.Errorf("invalid gzip file: %w", err)
-	}
-
-	dbPath := db.Path(cacheDir)
-	dbDir := filepath.Dir(dbPath)
-
-	if err = os.MkdirAll(dbDir, 0700); err != nil {
-		return xerrors.Errorf("failed to mkdir: %w", err)
-	}
-
-	file, err := os.Create(dbPath)
-	if err != nil {
-		return xerrors.Errorf("unable to open DB file: %w", err)
-	}
-	defer file.Close()
-
-	limited := io.LimitReader(gr, 2*gb)
-	if _, err = io.Copy(file, limited); err != nil {
-		return xerrors.Errorf("failed to save DB file: %w", err)
-	}
-
 	return nil
 }
 
-// UpdateMetadata updates the DB metadata
-func (c Client) UpdateMetadata(cacheDir string) error {
-	log.Logger.Debug("Updating database metadata...")
-
-	// make sure the DB has been successfully downloaded
-	if err := db.Init(cacheDir); err != nil {
-		return xerrors.Errorf("DB error: %w", err)
+func (c *Client) Clear(_ context.Context) error {
+	if err := os.RemoveAll(c.dbDir); err != nil {
+		return xerrors.Errorf("failed to remove vulnerability database: %w", err)
 	}
-	defer db.Close()
+	return nil
+}
 
-	metadata, err := c.dbc.GetMetadata()
+func (c *Client) updateDownloadedAt(ctx context.Context, dbDir string) error {
+	log.Debug("Updating database metadata...")
+
+	// We have to initialize a metadata client here
+	// since the destination may be different from the cache directory.
+	client := metadata.NewClient(dbDir)
+	meta, err := client.Get()
 	if err != nil {
 		return xerrors.Errorf("unable to get metadata: %w", err)
 	}
 
-	metadata.DownloadedAt = c.clock.Now().UTC()
-	if err = c.dbc.StoreMetadata(metadata, filepath.Join(cacheDir, "db")); err != nil {
-		return xerrors.Errorf("failed to store metadata: %w", err)
+	meta.DownloadedAt = clock.Now(ctx).UTC()
+	if err = client.Update(meta); err != nil {
+		return xerrors.Errorf("failed to update metadata: %w", err)
 	}
 
 	return nil
 }
 
-// Metadata defines the file meta
-type Metadata struct { // TODO: Move all Metadata things to trivy-db repo
-	fs       afero.Fs
-	filePath string
-}
-
-// NewMetadata is the factory method for file Metadata
-func NewMetadata(fs afero.Fs, cacheDir string) Metadata {
-	filePath := MetadataPath(cacheDir)
-	return Metadata{
-		fs:       fs,
-		filePath: filePath,
+func (c *Client) initArtifacts(opt types.RegistryOptions) oci.Artifacts {
+	if c.artifact != nil {
+		return oci.Artifacts{c.artifact}
 	}
+	return oci.NewArtifacts(c.dbRepositories, opt)
 }
 
-// MetadataPath returns the metaData file path
-func MetadataPath(cacheDir string) string {
-	dbPath := db.Path(cacheDir)
-	dbDir := filepath.Dir(dbPath)
-	return filepath.Join(dbDir, metadataFile)
-}
-
-// Delete deletes the file of database metadata
-func (m Metadata) Delete() error {
-	if err := m.fs.Remove(m.filePath); err != nil {
-		return xerrors.Errorf("unable to remove the metadata file: %w", err)
+func (c *Client) downloadDB(ctx context.Context, opt types.RegistryOptions, dst string) error {
+	log.InfoContext(ctx, "Downloading vulnerability DB...")
+	downloadOpt := oci.DownloadOption{
+		MediaType: dbMediaType,
+		Quiet:     c.quiet,
+	}
+	if err := c.initArtifacts(opt).Download(ctx, dst, downloadOpt); err != nil {
+		return xerrors.Errorf("failed to download vulnerability DB: %w", err)
 	}
 	return nil
 }
 
-// Get returns the file metadata
-func (m Metadata) Get() (db.Metadata, error) {
-	f, err := m.fs.Open(m.filePath)
+func (c *Client) ShowInfo() error {
+	meta, err := c.metadata.Get()
 	if err != nil {
-		return db.Metadata{}, xerrors.Errorf("unable to open a file: %w", err)
+		return xerrors.Errorf("something wrong with DB: %w", err)
 	}
-	defer f.Close()
-
-	var metadata db.Metadata
-	if err = json.NewDecoder(f).Decode(&metadata); err != nil {
-		return db.Metadata{}, xerrors.Errorf("unable to decode metadata: %w", err)
-	}
-	return metadata, nil
+	log.Debug("DB info", log.Int("schema", meta.Version), log.Time("updated_at", meta.UpdatedAt),
+		log.Time("next_update", meta.NextUpdate), log.Time("downloaded_at", meta.DownloadedAt))
+	return nil
 }
