@@ -1,367 +1,191 @@
-package db
+package db_test
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
-	"os"
+	"fmt"
 	"testing"
 	"time"
 
-	"github.com/aquasecurity/trivy/pkg/github"
-	"github.com/aquasecurity/trivy/pkg/indicator"
-
-	"github.com/spf13/afero"
-
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"golang.org/x/xerrors"
-
-	"k8s.io/utils/clock"
-	clocktesting "k8s.io/utils/clock/testing"
-
-	"github.com/aquasecurity/trivy-db/pkg/db"
-	"github.com/stretchr/testify/assert"
+	"github.com/aquasecurity/trivy-db/pkg/metadata"
+	"github.com/aquasecurity/trivy/internal/dbtest"
+	"github.com/aquasecurity/trivy/pkg/clock"
+	"github.com/aquasecurity/trivy/pkg/db"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
 )
 
 func TestClient_NeedsUpdate(t *testing.T) {
 	timeNextUpdateDay1 := time.Date(2019, 9, 1, 0, 0, 0, 0, time.UTC)
 	timeNextUpdateDay2 := time.Date(2019, 10, 2, 0, 0, 0, 0, time.UTC)
 
-	testCases := []struct {
-		name          string
-		light         bool
-		skip          bool
-		clock         clock.Clock
-		metadata      db.Metadata
-		expected      bool
-		expectedError error
+	tests := []struct {
+		name     string
+		skip     bool
+		metadata metadata.Metadata
+		want     bool
+		wantErr  string
 	}{
 		{
-			name:  "happy path",
-			light: false,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
-				Version:    1,
-				Type:       db.TypeFull,
+			name: "happy path",
+			metadata: metadata.Metadata{
+				Version:    db.SchemaVersion,
 				NextUpdate: timeNextUpdateDay1,
 			},
-			expected: true,
+			want: true,
 		},
 		{
 			name:     "happy path for first run",
-			light:    false,
-			clock:    clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{},
-			expected: true,
+			metadata: metadata.Metadata{},
+			want:     true,
 		},
 		{
-			name:  "happy path with different type",
-			light: true,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
-				Version:    1,
-				Type:       db.TypeFull,
-				NextUpdate: timeNextUpdateDay1,
-			},
-			expected: true,
-		},
-		{
-			name:  "happy path with old schema version",
-			light: true,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
+			name: "happy path with old schema version",
+			metadata: metadata.Metadata{
 				Version:    0,
-				Type:       db.TypeFull,
 				NextUpdate: timeNextUpdateDay1,
 			},
-			expected: true,
+			want: true,
 		},
 		{
-			name:  "happy path with --skip-update",
-			light: false,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
-				Version:    1,
-				Type:       db.TypeFull,
+			name: "happy path with --skip-update",
+			metadata: metadata.Metadata{
+				Version:    db.SchemaVersion,
 				NextUpdate: timeNextUpdateDay1,
 			},
+			skip: true,
+			want: false,
+		},
+		{
+			name: "skip downloading DB",
+			metadata: metadata.Metadata{
+				Version:    db.SchemaVersion,
+				NextUpdate: timeNextUpdateDay2,
+			},
+			want: false,
+		},
+		{
+			name: "newer schema version",
+			metadata: metadata.Metadata{
+				Version:    db.SchemaVersion + 1,
+				NextUpdate: timeNextUpdateDay2,
+			},
+			wantErr: fmt.Sprintf("the version of DB schema doesn't match. Local DB: %d, Expected: %d",
+				db.SchemaVersion+1, db.SchemaVersion),
+		},
+		{
+			name:     "--skip-update on the first run",
+			metadata: metadata.Metadata{},
 			skip:     true,
-			expected: false,
+			wantErr:  "--skip-update cannot be specified on the first run",
 		},
 		{
-			name:  "skip downloading DB",
-			light: false,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
-				Version:    1,
-				Type:       db.TypeFull,
-				NextUpdate: timeNextUpdateDay2,
-			},
-			expected: false,
-		},
-		{
-			name:  "newer schema version",
-			light: false,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
-				Version:    2,
-				Type:       db.TypeFull,
-				NextUpdate: timeNextUpdateDay2,
-			},
-			expectedError: xerrors.New("the version of DB schema doesn't match. Local DB: 2, Expected: 1"),
-		},
-		{
-			name:          "--skip-update on the first run",
-			light:         false,
-			clock:         clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata:      db.Metadata{},
-			skip:          true,
-			expectedError: xerrors.New("--skip-update cannot be specified on the first run"),
-		},
-		{
-			name:  "--skip-update with different schema version",
-			light: false,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
+			name: "--skip-update with different schema version",
+			metadata: metadata.Metadata{
 				Version:    0,
-				Type:       db.TypeFull,
 				NextUpdate: timeNextUpdateDay1,
 			},
-			skip:          true,
-			expectedError: xerrors.New("--skip-update cannot be specified with the old DB"),
+			skip: true,
+			wantErr: fmt.Sprintf("--skip-update cannot be specified with the old DB schema. Local DB: %d, Expected: %d",
+				0, db.SchemaVersion),
 		},
 		{
-			name:  "happy with old DownloadedAt",
-			light: false,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
-				Version:      1,
-				Type:         db.TypeFull,
+			name: "happy with old DownloadedAt",
+			metadata: metadata.Metadata{
+				Version:      db.SchemaVersion,
 				NextUpdate:   timeNextUpdateDay1,
 				DownloadedAt: time.Date(2019, 9, 30, 22, 30, 0, 0, time.UTC),
 			},
-			expected: true,
+			want: true,
 		},
 		{
-			name:  "skip downloading DB with recent DownloadedAt",
-			light: false,
-			clock: clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC)),
-			metadata: db.Metadata{
-				Version:      1,
-				Type:         db.TypeFull,
+			name: "skip downloading DB with recent DownloadedAt",
+			metadata: metadata.Metadata{
+				Version:      db.SchemaVersion,
 				NextUpdate:   timeNextUpdateDay1,
 				DownloadedAt: time.Date(2019, 9, 30, 23, 30, 0, 0, time.UTC),
 			},
-			expected: false,
+			want: false,
 		},
 	}
 
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			fs := afero.NewMemMapFs()
-			metadata := NewMetadata(fs, "/cache")
-			if tc.metadata != (db.Metadata{}) {
-				b, err := json.Marshal(tc.metadata)
-				require.NoError(t, err)
-				err = afero.WriteFile(fs, metadata.filePath, b, 0600)
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			dbDir := db.Dir(t.TempDir())
+			if tt.metadata != (metadata.Metadata{}) {
+				meta := metadata.NewClient(dbDir)
+				err := meta.Update(tt.metadata)
 				require.NoError(t, err)
 			}
 
-			client := Client{
-				clock:    tc.clock,
-				metadata: metadata,
-			}
+			// Set a fake time
+			ctx := clock.With(context.Background(), time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC))
 
-			needsUpdate, err := client.NeedsUpdate("test", tc.light, tc.skip)
+			client := db.NewClient(dbDir, true)
+			needsUpdate, err := client.NeedsUpdate(ctx, "test", tt.skip)
 
 			switch {
-			case tc.expectedError != nil:
-				assert.EqualError(t, err, tc.expectedError.Error(), tc.name)
+			case tt.wantErr != "":
+				require.Error(t, err)
+				assert.Contains(t, err.Error(), tt.wantErr, tt.name)
 			default:
-				assert.NoError(t, err, tc.name)
+				require.NoError(t, err, tt.name)
 			}
 
-			assert.Equal(t, tc.expected, needsUpdate)
+			assert.Equal(t, tt.want, needsUpdate)
 		})
 	}
 }
 
 func TestClient_Download(t *testing.T) {
-	testCases := []struct {
-		name            string
-		light           bool
-		downloadDB      []github.DownloadDBExpectation
-		expectedContent []byte
-		expectedError   error
+
+	tests := []struct {
+		name    string
+		input   string
+		want    metadata.Metadata
+		wantErr string
 	}{
 		{
 			name:  "happy path",
-			light: false,
-			downloadDB: []github.DownloadDBExpectation{
-				{
-					Args: github.DownloadDBInput{FileName: fullDB},
-					ReturnArgs: github.DownloadDBOutput{
-						FileName: "testdata/test.db.gz",
-					},
-				},
+			input: "testdata/db.tar.gz",
+			want: metadata.Metadata{
+				Version:      1,
+				NextUpdate:   time.Date(3000, 1, 1, 18, 5, 43, 198355188, time.UTC),
+				UpdatedAt:    time.Date(3000, 1, 1, 12, 5, 43, 198355588, time.UTC),
+				DownloadedAt: time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC),
 			},
 		},
 		{
-			name:  "DownloadDB returns an error",
-			light: false,
-			downloadDB: []github.DownloadDBExpectation{
-				{
-					Args: github.DownloadDBInput{FileName: fullDB},
-					ReturnArgs: github.DownloadDBOutput{
-						Err: xerrors.New("download failed"),
-					},
-				},
-			},
-			expectedError: xerrors.New("failed to download vulnerability DB: download failed"),
-		},
-		{
-			name:  "invalid gzip",
-			light: false,
-			downloadDB: []github.DownloadDBExpectation{
-				{
-					Args: github.DownloadDBInput{FileName: fullDB},
-					ReturnArgs: github.DownloadDBOutput{
-						FileName: "testdata/invalid.db.gz",
-					},
-				},
-			},
-			expectedError: xerrors.New("invalid gzip file: unexpected EOF"),
+			name:    "invalid gzip",
+			input:   "testdata/trivy.db",
+			wantErr: "OCI artifact error: failed to download vulnerability DB",
 		},
 	}
 
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			mockConfig := new(mockDbOperation)
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// Set a fake time
+			ctx := clock.With(context.Background(), time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC))
 
-			mockGitHubClient, err := github.NewMockClient(tc.downloadDB)
-			require.NoError(t, err, tc.name)
+			// Fake DB
+			art := dbtest.NewFakeDB(t, tt.input, dbtest.FakeDBOptions{})
 
-			fs := afero.NewMemMapFs()
-			metadata := NewMetadata(fs, "/cache")
-
-			dir, err := os.MkdirTemp("", "db")
-			require.NoError(t, err, tc.name)
-			defer os.RemoveAll(dir)
-
-			pb := indicator.NewProgressBar(true)
-			client := NewClient(mockConfig, mockGitHubClient, pb, nil, metadata)
-			ctx := context.Background()
-			err = client.Download(ctx, dir, tc.light)
-
-			switch {
-			case tc.expectedError != nil:
-				assert.EqualError(t, err, tc.expectedError.Error(), tc.name)
-			default:
-				assert.NoError(t, err, tc.name)
+			dbDir := db.Dir(t.TempDir())
+			client := db.NewClient(dbDir, true, db.WithOCIArtifact(art))
+			err := client.Download(ctx, dbDir, ftypes.RegistryOptions{})
+			if tt.wantErr != "" {
+				require.Error(t, err)
+				assert.ErrorContains(t, err, tt.wantErr)
+				return
 			}
+			require.NoError(t, err)
 
-			mockGitHubClient.AssertExpectations(t)
-		})
-	}
-}
+			meta := metadata.NewClient(dbDir)
+			got, err := meta.Get()
+			require.NoError(t, err)
 
-func TestClient_UpdateMetadata(t *testing.T) {
-	timeDownloadedAt := clocktesting.NewFakeClock(time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC))
-	testCases := []struct {
-		name                     string
-		clock                    clock.Clock
-		getMetadataExpectation   dbOperationGetMetadataExpectation
-		storeMetadataExpectation dbOperationStoreMetadataExpectation
-		expectedError            error
-	}{
-		{
-			name:  "happy path",
-			clock: timeDownloadedAt,
-			getMetadataExpectation: dbOperationGetMetadataExpectation{
-				Returns: dbOperationGetMetadataReturns{
-					Metadata: db.Metadata{
-						Version:    1,
-						Type:       1,
-						NextUpdate: time.Date(2020, 4, 30, 23, 59, 59, 0, time.UTC),
-						UpdatedAt:  time.Date(2006, 4, 30, 23, 59, 59, 0, time.UTC),
-					},
-					Err: nil,
-				},
-			},
-			storeMetadataExpectation: dbOperationStoreMetadataExpectation{
-				Metadata: db.Metadata{
-					Version:      1,
-					Type:         1,
-					NextUpdate:   time.Date(2020, 4, 30, 23, 59, 59, 0, time.UTC),
-					UpdatedAt:    time.Date(2006, 4, 30, 23, 59, 59, 0, time.UTC),
-					DownloadedAt: timeDownloadedAt.Now(),
-				},
-			},
-		},
-		{
-			name:  "sad path, get metadata fails",
-			clock: timeDownloadedAt,
-			getMetadataExpectation: dbOperationGetMetadataExpectation{
-				Returns: dbOperationGetMetadataReturns{
-					Err: errors.New("get metadata failed"),
-				},
-			},
-			expectedError: errors.New("unable to get metadata: get metadata failed"),
-		},
-		{
-			name:  "sad path, store metadata fails",
-			clock: timeDownloadedAt,
-			getMetadataExpectation: dbOperationGetMetadataExpectation{
-				Returns: dbOperationGetMetadataReturns{
-					Metadata: db.Metadata{
-						Version:    1,
-						Type:       1,
-						NextUpdate: time.Date(2020, 4, 30, 23, 59, 59, 0, time.UTC),
-						UpdatedAt:  time.Date(2006, 4, 30, 23, 59, 59, 0, time.UTC),
-					},
-					Err: nil,
-				},
-			},
-			storeMetadataExpectation: dbOperationStoreMetadataExpectation{
-				Metadata: db.Metadata{
-					Version:      1,
-					Type:         1,
-					NextUpdate:   time.Date(2020, 4, 30, 23, 59, 59, 0, time.UTC),
-					UpdatedAt:    time.Date(2006, 4, 30, 23, 59, 59, 0, time.UTC),
-					DownloadedAt: timeDownloadedAt.Now(),
-				},
-				Returns: dbOperationStoreMetadataReturns{
-					Err: errors.New("store metadata failed"),
-				},
-			},
-			expectedError: errors.New("failed to store metadata: store metadata failed"),
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			mockConfig := new(mockDbOperation)
-			mockConfig.ApplyGetMetadataExpectation(tc.getMetadataExpectation)
-			mockConfig.ApplyStoreMetadataExpectation(tc.storeMetadataExpectation)
-
-			fs := afero.NewMemMapFs()
-			metadata := NewMetadata(fs, "/cache")
-
-			dir, err := os.MkdirTemp("", "db")
-			require.NoError(t, err, tc.name)
-			defer os.RemoveAll(dir)
-
-			pb := indicator.NewProgressBar(true)
-			client := NewClient(mockConfig, nil, pb, tc.clock, metadata)
-
-			err = client.UpdateMetadata(dir)
-			switch {
-			case tc.expectedError != nil:
-				assert.EqualError(t, err, tc.expectedError.Error(), tc.name)
-			default:
-				assert.NoError(t, err, tc.name)
-			}
+			assert.Equal(t, tt.want, got)
 		})
 	}
 }
