@@ -1,20 +1,23 @@
 package library
 
 import (
+	"context"
+
 	"golang.org/x/xerrors"
 
-	ftypes "github.com/aquasecurity/fanal/types"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
+	"github.com/aquasecurity/trivy/pkg/log"
 	"github.com/aquasecurity/trivy/pkg/types"
 )
 
-// Detect scans and returns vulnerabilities of library
-func Detect(libType string, pkgs []ftypes.Package) ([]types.DetectedVulnerability, error) {
-	driver, err := NewDriver(libType)
-	if err != nil {
-		return nil, xerrors.Errorf("failed to new driver: %w", err)
+// Detect scans language-specific packages and returns vulnerabilities.
+func Detect(ctx context.Context, libType ftypes.LangType, pkgs []ftypes.Package) ([]types.DetectedVulnerability, error) {
+	driver, ok := NewDriver(libType)
+	if !ok {
+		return nil, nil
 	}
 
-	vulns, err := detect(driver, pkgs)
+	vulns, err := detect(ctx, driver, pkgs)
 	if err != nil {
 		return nil, xerrors.Errorf("failed to scan %s vulnerabilities: %w", driver.Type(), err)
 	}
@@ -22,17 +25,23 @@ func Detect(libType string, pkgs []ftypes.Package) ([]types.DetectedVulnerabilit
 	return vulns, nil
 }
 
-func detect(driver Driver, libs []ftypes.Package) ([]types.DetectedVulnerability, error) {
+func detect(ctx context.Context, driver Driver, pkgs []ftypes.Package) ([]types.DetectedVulnerability, error) {
 	var vulnerabilities []types.DetectedVulnerability
-	for _, lib := range libs {
-		vulns, err := driver.Detect(lib.Name, lib.Version)
+	for _, pkg := range pkgs {
+		if pkg.Version == "" {
+			log.DebugContext(ctx, "Skipping vulnerability scan as no version is detected for the package",
+				log.String("name", pkg.Name))
+			continue
+		}
+		vulns, err := driver.DetectVulnerabilities(pkg.ID, pkg.Name, pkg.Version)
 		if err != nil {
 			return nil, xerrors.Errorf("failed to detect %s vulnerabilities: %w", driver.Type(), err)
 		}
 
 		for i := range vulns {
-			vulns[i].Layer = lib.Layer
-			vulns[i].PkgPath = lib.FilePath
+			vulns[i].Layer = pkg.Layer
+			vulns[i].PkgPath = pkg.FilePath
+			vulns[i].PkgIdentifier = pkg.Identifier
 		}
 		vulnerabilities = append(vulnerabilities, vulns...)
 	}
