@@ -1,21 +1,26 @@
 package ospkg
 
 import (
+	"context"
 	"time"
 
-	"github.com/google/wire"
+	"github.com/samber/lo"
 	"golang.org/x/xerrors"
 
-	fos "github.com/aquasecurity/fanal/analyzer/os"
-	ftypes "github.com/aquasecurity/fanal/types"
+	"github.com/aquasecurity/trivy/pkg/detector/ospkg/alma"
 	"github.com/aquasecurity/trivy/pkg/detector/ospkg/alpine"
 	"github.com/aquasecurity/trivy/pkg/detector/ospkg/amazon"
+	"github.com/aquasecurity/trivy/pkg/detector/ospkg/azure"
+	"github.com/aquasecurity/trivy/pkg/detector/ospkg/chainguard"
 	"github.com/aquasecurity/trivy/pkg/detector/ospkg/debian"
 	"github.com/aquasecurity/trivy/pkg/detector/ospkg/oracle"
 	"github.com/aquasecurity/trivy/pkg/detector/ospkg/photon"
 	"github.com/aquasecurity/trivy/pkg/detector/ospkg/redhat"
+	"github.com/aquasecurity/trivy/pkg/detector/ospkg/rocky"
 	"github.com/aquasecurity/trivy/pkg/detector/ospkg/suse"
 	"github.com/aquasecurity/trivy/pkg/detector/ospkg/ubuntu"
+	"github.com/aquasecurity/trivy/pkg/detector/ospkg/wolfi"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
 	"github.com/aquasecurity/trivy/pkg/log"
 	"github.com/aquasecurity/trivy/pkg/types"
 )
@@ -24,37 +29,56 @@ var (
 	// ErrUnsupportedOS defines error for unsupported OS
 	ErrUnsupportedOS = xerrors.New("unsupported os")
 
-	// SuperSet binds dependencies for OS scan
-	SuperSet = wire.NewSet(
-		wire.Struct(new(Detector)),
-		wire.Bind(new(Operation), new(Detector)),
-	)
+	drivers = map[ftypes.OSType]Driver{
+		ftypes.Alpine:             alpine.NewScanner(),
+		ftypes.Alma:               alma.NewScanner(),
+		ftypes.Amazon:             amazon.NewScanner(),
+		ftypes.Azure:              azure.NewAzureScanner(),
+		ftypes.CBLMariner:         azure.NewMarinerScanner(),
+		ftypes.Debian:             debian.NewScanner(),
+		ftypes.Ubuntu:             ubuntu.NewScanner(),
+		ftypes.RedHat:             redhat.NewScanner(),
+		ftypes.CentOS:             redhat.NewScanner(),
+		ftypes.Rocky:              rocky.NewScanner(),
+		ftypes.Oracle:             oracle.NewScanner(),
+		ftypes.OpenSUSETumbleweed: suse.NewScanner(suse.OpenSUSETumbleweed),
+		ftypes.OpenSUSELeap:       suse.NewScanner(suse.OpenSUSE),
+		ftypes.SLES:               suse.NewScanner(suse.SUSEEnterpriseLinux),
+		ftypes.SLEMicro:           suse.NewScanner(suse.SUSEEnterpriseLinuxMicro),
+		ftypes.Photon:             photon.NewScanner(),
+		ftypes.Wolfi:              wolfi.NewScanner(),
+		ftypes.Chainguard:         chainguard.NewScanner(),
+	}
 )
 
-// Operation defines operation of OSpkg scan
-type Operation interface {
-	Detect(string, string, string, time.Time, []ftypes.Package) ([]types.DetectedVulnerability, bool, error)
+// RegisterDriver is defined for extensibility and not supposed to be used in Trivy.
+func RegisterDriver(name ftypes.OSType, driver Driver) {
+	drivers[name] = driver
 }
 
 // Driver defines operations for OS package scan
 type Driver interface {
-	Detect(string, []ftypes.Package) ([]types.DetectedVulnerability, error)
-	IsSupportedVersion(string, string) bool
+	Detect(context.Context, string, *ftypes.Repository, []ftypes.Package) ([]types.DetectedVulnerability, error)
+	IsSupportedVersion(context.Context, ftypes.OSType, string) bool
 }
 
-// Detector implements Operation
-type Detector struct{}
-
 // Detect detects the vulnerabilities
-func (d Detector) Detect(_, osFamily, osName string, _ time.Time, pkgs []ftypes.Package) ([]types.DetectedVulnerability, bool, error) {
-	driver := newDriver(osFamily, osName)
-	if driver == nil {
+func Detect(ctx context.Context, _, osFamily ftypes.OSType, osName string, repo *ftypes.Repository, _ time.Time, pkgs []ftypes.Package) ([]types.DetectedVulnerability, bool, error) {
+	ctx = log.WithContextPrefix(ctx, string(osFamily))
+
+	driver, err := newDriver(osFamily)
+	if err != nil {
 		return nil, false, ErrUnsupportedOS
 	}
 
-	eosl := !driver.IsSupportedVersion(osFamily, osName)
+	eosl := !driver.IsSupportedVersion(ctx, osFamily, osName)
 
-	vulns, err := driver.Detect(osName, pkgs)
+	// Package `gpg-pubkey` doesn't use the correct version.
+	// We don't need to find vulnerabilities for this package.
+	filteredPkgs := lo.Filter(pkgs, func(pkg ftypes.Package, index int) bool {
+		return pkg.Name != "gpg-pubkey"
+	})
+	vulns, err := driver.Detect(ctx, osName, repo, filteredPkgs)
 	if err != nil {
 		return nil, false, xerrors.Errorf("failed detection: %w", err)
 	}
@@ -62,31 +86,11 @@ func (d Detector) Detect(_, osFamily, osName string, _ time.Time, pkgs []ftypes.
 	return vulns, eosl, nil
 }
 
-// nolint: gocyclo
-// TODO: fix cyclometic complexity by removing default
-func newDriver(osFamily, osName string) Driver {
-	// TODO: use DI and change struct names
-	switch osFamily {
-	case fos.Alpine:
-		return alpine.NewScanner()
-	case fos.Debian:
-		return debian.NewScanner()
-	case fos.Ubuntu:
-		return ubuntu.NewScanner()
-	case fos.RedHat, fos.CentOS:
-		return redhat.NewScanner()
-	case fos.Amazon:
-		return amazon.NewScanner()
-	case fos.Oracle:
-		return oracle.NewScanner()
-	case fos.OpenSUSELeap:
-		return suse.NewScanner(suse.OpenSUSE)
-	case fos.SLES:
-		return suse.NewScanner(suse.SUSEEnterpriseLinux)
-	case fos.Photon:
-		return photon.NewScanner()
-	default:
-		log.Logger.Warnf("unsupported os : %s", osFamily)
-		return nil
+func newDriver(osFamily ftypes.OSType) (Driver, error) {
+	if driver, ok := drivers[osFamily]; ok {
+		return driver, nil
 	}
+
+	log.Warn("Unsupported os", log.String("family", string(osFamily)))
+	return nil, ErrUnsupportedOS
 }
