@@ -1,20 +1,30 @@
 package photon
 
 import (
+	"context"
+	"time"
+
 	version "github.com/knqyf263/go-rpm-version"
 	"golang.org/x/xerrors"
 
-	ftypes "github.com/aquasecurity/fanal/types"
 	"github.com/aquasecurity/trivy-db/pkg/vulnsrc/photon"
+	osver "github.com/aquasecurity/trivy/pkg/detector/ospkg/version"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
 	"github.com/aquasecurity/trivy/pkg/log"
 	"github.com/aquasecurity/trivy/pkg/scanner/utils"
 	"github.com/aquasecurity/trivy/pkg/types"
 )
 
-// EOL can't be found for photon https://github.com/vmware/photon/issues/1031
-//var (
-//	eolDates = map[string]time.Time{}
-//)
+var (
+	eolDates = map[string]time.Time{
+		"1.0": time.Date(2022, 2, 28, 23, 59, 59, 0, time.UTC),
+		"2.0": time.Date(2022, 12, 31, 23, 59, 59, 0, time.UTC),
+		// The following versions don't have the EOL dates yet.
+		// See https://blogs.vmware.com/vsphere/2022/01/photon-1-x-end-of-support-announcement.html
+		"3.0": time.Date(2024, 6, 30, 23, 59, 59, 0, time.UTC),
+		"4.0": time.Date(2025, 12, 31, 23, 59, 59, 0, time.UTC),
+	}
+)
 
 // Scanner implements the Photon scanner
 type Scanner struct {
@@ -29,10 +39,9 @@ func NewScanner() *Scanner {
 }
 
 // Detect scans and returns vulnerabilities using photon scanner
-func (s *Scanner) Detect(osVer string, pkgs []ftypes.Package) ([]types.DetectedVulnerability, error) {
-	log.Logger.Info("Detecting Photon Linux vulnerabilities...")
-	log.Logger.Debugf("Photon Linux: os version: %s", osVer)
-	log.Logger.Debugf("Photon Linux: the number of packages: %d", len(pkgs))
+func (s *Scanner) Detect(ctx context.Context, osVer string, _ *ftypes.Repository, pkgs []ftypes.Package) ([]types.DetectedVulnerability, error) {
+	log.InfoContext(ctx, "Detecting vulnerabilities...", log.String("os_version", osVer),
+		log.Int("pkg_num", len(pkgs)))
 
 	var vulns []types.DetectedVulnerability
 	for _, pkg := range pkgs {
@@ -47,10 +56,13 @@ func (s *Scanner) Detect(osVer string, pkgs []ftypes.Package) ([]types.DetectedV
 			fixedVersion := version.NewVersion(adv.FixedVersion)
 			vuln := types.DetectedVulnerability{
 				VulnerabilityID:  adv.VulnerabilityID,
+				PkgID:            pkg.ID,
 				PkgName:          pkg.Name,
 				InstalledVersion: installed,
+				PkgIdentifier:    pkg.Identifier,
 				Layer:            pkg.Layer,
 				Custom:           adv.Custom,
+				DataSource:       adv.DataSource,
 			}
 			if installedVersion.LessThan(fixedVersion) {
 				vuln.FixedVersion = adv.FixedVersion
@@ -61,7 +73,7 @@ func (s *Scanner) Detect(osVer string, pkgs []ftypes.Package) ([]types.DetectedV
 	return vulns, nil
 }
 
-// IsSupportedVersion checks is OSFamily can be scanned
-func (s *Scanner) IsSupportedVersion(_, _ string) bool {
-	return true
+// IsSupportedVersion checks if the version is supported.
+func (s *Scanner) IsSupportedVersion(ctx context.Context, osFamily ftypes.OSType, osVer string) bool {
+	return osver.Supported(ctx, eolDates, osFamily, osVer)
 }
