@@ -1,22 +1,45 @@
 package redhat
 
 import (
+	"context"
+	"fmt"
+	"slices"
+	"sort"
 	"strings"
 	"time"
 
 	version "github.com/knqyf263/go-rpm-version"
 	"golang.org/x/xerrors"
-	"k8s.io/utils/clock"
 
-	"github.com/aquasecurity/fanal/analyzer/os"
-	ftypes "github.com/aquasecurity/fanal/types"
-	"github.com/aquasecurity/trivy-db/pkg/vulnsrc/redhat"
+	dbTypes "github.com/aquasecurity/trivy-db/pkg/types"
+	redhat "github.com/aquasecurity/trivy-db/pkg/vulnsrc/redhat-oval"
+	"github.com/aquasecurity/trivy-db/pkg/vulnsrc/vulnerability"
+	osver "github.com/aquasecurity/trivy/pkg/detector/ospkg/version"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
 	"github.com/aquasecurity/trivy/pkg/log"
 	"github.com/aquasecurity/trivy/pkg/scanner/utils"
 	"github.com/aquasecurity/trivy/pkg/types"
 )
 
 var (
+	defaultContentSets = map[string][]string{
+		"6": {
+			"rhel-6-server-rpms",
+			"rhel-6-server-extras-rpms",
+		},
+		"7": {
+			"rhel-7-server-rpms",
+			"rhel-7-server-extras-rpms",
+		},
+		"8": {
+			"rhel-8-for-x86_64-baseos-rpms",
+			"rhel-8-for-x86_64-appstream-rpms",
+		},
+		"9": {
+			"rhel-9-for-x86_64-baseos-rpms",
+			"rhel-9-for-x86_64-appstream-rpms",
+		},
+	}
 	redhatEOLDates = map[string]time.Time{
 		"4": time.Date(2017, 5, 31, 23, 59, 59, 0, time.UTC),
 		"5": time.Date(2020, 11, 30, 23, 59, 59, 0, time.UTC),
@@ -24,6 +47,7 @@ var (
 		// N/A
 		"7": time.Date(3000, 1, 1, 23, 59, 59, 0, time.UTC),
 		"8": time.Date(3000, 1, 1, 23, 59, 59, 0, time.UTC),
+		"9": time.Date(3000, 1, 1, 23, 59, 59, 0, time.UTC),
 	}
 	centosEOLDates = map[string]time.Time{
 		"3": time.Date(2010, 10, 31, 23, 59, 59, 0, time.UTC),
@@ -38,127 +62,122 @@ var (
 	}
 )
 
-type options struct {
-	clock clock.Clock
-}
-
-type option func(*options)
-
-func WithClock(clock clock.Clock) option {
-	return func(opts *options) {
-		opts.clock = clock
-	}
-}
-
-// Scanner implements the Alpine scanner
+// Scanner implements the RedHat scanner
 type Scanner struct {
 	vs redhat.VulnSrc
-	*options
 }
 
 // NewScanner is the factory method for Scanner
-func NewScanner(opts ...option) *Scanner {
-	o := &options{
-		clock: clock.RealClock{},
-	}
-
-	for _, opt := range opts {
-		opt(o)
-	}
+func NewScanner() *Scanner {
 	return &Scanner{
-		vs:      redhat.NewVulnSrc(),
-		options: o,
+		vs: redhat.NewVulnSrc(),
 	}
 }
 
-// Detect scans and returns redhat vulenrabilities
-func (s *Scanner) Detect(osVer string, pkgs []ftypes.Package) ([]types.DetectedVulnerability, error) {
-	log.Logger.Info("Detecting RHEL/CentOS vulnerabilities...")
-	if strings.Count(osVer, ".") > 0 {
-		osVer = osVer[:strings.Index(osVer, ".")]
-	}
-	log.Logger.Debugf("redhat: os version: %s", osVer)
-	log.Logger.Debugf("redhat: the number of packages: %d", len(pkgs))
+// Detect scans and returns redhat vulnerabilities
+func (s *Scanner) Detect(ctx context.Context, osVer string, _ *ftypes.Repository, pkgs []ftypes.Package) ([]types.DetectedVulnerability, error) {
+	osVer = osver.Major(osVer)
+	log.InfoContext(ctx, "Detecting RHEL/CentOS vulnerabilities...", log.String("os_version", osVer),
+		log.Int("pkg_num", len(pkgs)))
 
 	var vulns []types.DetectedVulnerability
 	for _, pkg := range pkgs {
-		if !s.isFromSupportedVendor(pkg) {
-			log.Logger.Debugf("Skipping %s: unsupported vendor", pkg.Name)
+		if !isFromSupportedVendor(pkg) {
+			log.DebugContext(ctx, "Skipping the package with unsupported vendor", log.String("package", pkg.Name))
 			continue
 		}
 
-		// For Red Hat Security Data API containing only source package names
-		pkgName := addModularNamespace(pkg.SrcName, pkg.Modularitylabel)
-		advisories, err := s.vs.Get(osVer, pkgName)
+		detectedVulns, err := s.detect(osVer, pkg)
 		if err != nil {
-			return nil, xerrors.Errorf("failed to get Red Hat advisories: %w", err)
+			return nil, xerrors.Errorf("redhat vulnerability detection error: %w", err)
 		}
-
-		installed := utils.FormatVersion(pkg)
-		installedVersion := version.NewVersion(installed)
-
-		for _, adv := range advisories {
-			if adv.FixedVersion != "" {
-				continue
-			}
-			vuln := types.DetectedVulnerability{
-				VulnerabilityID:  adv.VulnerabilityID,
-				PkgName:          pkg.Name,
-				InstalledVersion: installed,
-				Layer:            pkg.Layer,
-				Custom:           adv.Custom,
-			}
-			vulns = append(vulns, vuln)
-		}
-
-		// For Red Hat OVAL v2 containing only binary package names
-		pkgName = addModularNamespace(pkg.Name, pkg.Modularitylabel)
-		advisories, err = s.vs.Get(osVer, pkgName)
-		if err != nil {
-			return nil, xerrors.Errorf("failed to get Red Hat advisories: %w", err)
-		}
-
-		for _, adv := range advisories {
-			fixedVersion := version.NewVersion(adv.FixedVersion)
-			if installedVersion.LessThan(fixedVersion) {
-				vuln := types.DetectedVulnerability{
-					VulnerabilityID:  adv.VulnerabilityID,
-					PkgName:          pkg.Name,
-					InstalledVersion: installed,
-					FixedVersion:     fixedVersion.String(),
-					Layer:            pkg.Layer,
-				}
-				vulns = append(vulns, vuln)
-			}
-		}
+		vulns = append(vulns, detectedVulns...)
 	}
 	return vulns, nil
 }
 
-// IsSupportedVersion checks is OSFamily can be scanned with Redhat scanner
-func (s *Scanner) IsSupportedVersion(osFamily, osVer string) bool {
-	if strings.Count(osVer, ".") > 0 {
-		osVer = osVer[:strings.Index(osVer, ".")]
+func (s *Scanner) detect(osVer string, pkg ftypes.Package) ([]types.DetectedVulnerability, error) {
+	// For Red Hat OVAL v2 containing only binary package names
+	pkgName := addModularNamespace(pkg.Name, pkg.Modularitylabel)
+
+	var contentSets []string
+	var nvr string
+	if pkg.BuildInfo == nil {
+		contentSets = defaultContentSets[osVer]
+	} else {
+		contentSets = pkg.BuildInfo.ContentSets
+		nvr = fmt.Sprintf("%s-%s", pkg.BuildInfo.Nvr, pkg.BuildInfo.Arch)
 	}
 
-	var eolDate time.Time
-	var ok bool
-	if osFamily == os.RedHat {
-		eolDate, ok = redhatEOLDates[osVer]
-	} else if osFamily == os.CentOS {
-		eolDate, ok = centosEOLDates[osVer]
-	}
-	if !ok {
-		log.Logger.Warnf("This OS version is not on the EOL list: %s %s", osFamily, osVer)
-		return false
+	advisories, err := s.vs.Get(pkgName, contentSets, []string{nvr})
+	if err != nil {
+		return nil, xerrors.Errorf("failed to get Red Hat advisories: %w", err)
 	}
 
-	return s.clock.Now().Before(eolDate)
+	// Choose the latest fixed version for each CVE-ID (empty for unpatched vulns).
+	// Take the single RHSA-ID with the latest fixed version (for patched vulns).
+	uniqAdvisories := make(map[string]dbTypes.Advisory)
+	for _, adv := range advisories {
+		// If Arches for advisory are empty or pkg.Arch is "noarch", then any Arches are affected
+		if len(adv.Arches) != 0 && pkg.Arch != "noarch" {
+			if !slices.Contains(adv.Arches, pkg.Arch) {
+				continue
+			}
+		}
+
+		if a, ok := uniqAdvisories[adv.VulnerabilityID]; ok {
+			if version.NewVersion(a.FixedVersion).LessThan(version.NewVersion(adv.FixedVersion)) {
+				uniqAdvisories[adv.VulnerabilityID] = adv
+			}
+		} else {
+			uniqAdvisories[adv.VulnerabilityID] = adv
+		}
+	}
+
+	var vulns []types.DetectedVulnerability
+	for _, adv := range uniqAdvisories {
+		vuln := types.DetectedVulnerability{
+			VulnerabilityID:  adv.VulnerabilityID,
+			VendorIDs:        adv.VendorIDs, // Will be empty for unpatched vulnerabilities
+			PkgID:            pkg.ID,
+			PkgName:          pkg.Name,
+			InstalledVersion: utils.FormatVersion(pkg),
+			FixedVersion:     version.NewVersion(adv.FixedVersion).String(), // Will be empty for unpatched vulnerabilities
+			PkgIdentifier:    pkg.Identifier,
+			Status:           adv.Status,
+			Layer:            pkg.Layer,
+			SeveritySource:   vulnerability.RedHat,
+			Vulnerability: dbTypes.Vulnerability{
+				Severity: adv.Severity.String(),
+			},
+			Custom: adv.Custom,
+		}
+
+		// Keep unpatched and affected vulnerabilities
+		if adv.FixedVersion == "" || version.NewVersion(vuln.InstalledVersion).LessThan(version.NewVersion(adv.FixedVersion)) {
+			vulns = append(vulns, vuln)
+		}
+	}
+
+	sort.Slice(vulns, func(i, j int) bool {
+		return vulns[i].VulnerabilityID < vulns[j].VulnerabilityID
+	})
+	return vulns, nil
 }
 
-func (s *Scanner) isFromSupportedVendor(pkg ftypes.Package) bool {
-	for _, s := range excludedVendorsSuffix {
-		if strings.HasSuffix(pkg.Release, s) {
+// IsSupportedVersion checks is OSFamily can be scanned with Redhat scanner
+func (s *Scanner) IsSupportedVersion(ctx context.Context, osFamily ftypes.OSType, osVer string) bool {
+	osVer = osver.Major(osVer)
+	if osFamily == ftypes.CentOS {
+		return osver.Supported(ctx, centosEOLDates, osFamily, osVer)
+	}
+
+	return osver.Supported(ctx, redhatEOLDates, osFamily, osVer)
+}
+
+func isFromSupportedVendor(pkg ftypes.Package) bool {
+	for _, suffix := range excludedVendorsSuffix {
+		if strings.HasSuffix(pkg.Release, suffix) {
 			return false
 		}
 	}
