@@ -2,12 +2,15 @@ package report_test
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	dbTypes "github.com/aquasecurity/trivy-db/pkg/types"
+	"github.com/aquasecurity/trivy-db/pkg/vulnsrc/vulnerability"
 	"github.com/aquasecurity/trivy/pkg/report"
 	"github.com/aquasecurity/trivy/pkg/types"
 )
@@ -16,7 +19,7 @@ func TestReportWriter_JSON(t *testing.T) {
 	testCases := []struct {
 		name          string
 		detectedVulns []types.DetectedVulnerability
-		want          report.Report
+		want          types.Report
 	}{
 		{
 			name: "happy path",
@@ -31,14 +34,17 @@ func TestReportWriter_JSON(t *testing.T) {
 						Title:       "foobar",
 						Description: "baz",
 						Severity:    "HIGH",
+						VendorSeverity: map[dbTypes.SourceID]dbTypes.Severity{
+							vulnerability.NVD: dbTypes.SeverityHigh,
+						},
 					},
 				},
 			},
-			want: report.Report{
+			want: types.Report{
 				SchemaVersion: 2,
 				ArtifactName:  "alpine:3.14",
-				Results: report.Results{
-					report.Result{
+				Results: types.Results{
+					types.Result{
 						Target: "foojson",
 						Vulnerabilities: []types.DetectedVulnerability{
 							{
@@ -51,6 +57,9 @@ func TestReportWriter_JSON(t *testing.T) {
 									Title:       "foobar",
 									Description: "baz",
 									Severity:    "HIGH",
+									VendorSeverity: map[dbTypes.SourceID]dbTypes.Severity{
+										vulnerability.NVD: dbTypes.SeverityHigh,
+									},
 								},
 							},
 						},
@@ -62,14 +71,15 @@ func TestReportWriter_JSON(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			jw := report.JSONWriter{}
-			jsonWritten := bytes.Buffer{}
-			jw.Output = &jsonWritten
+			jsonWritten := bytes.NewBuffer(nil)
+			jw := report.JSONWriter{
+				Output: jsonWritten,
+			}
 
-			inputResults := report.Report{
+			inputResults := types.Report{
 				SchemaVersion: 2,
 				ArtifactName:  "alpine:3.14",
-				Results: report.Results{
+				Results: types.Results{
 					{
 						Target:          "foojson",
 						Vulnerabilities: tc.detectedVulns,
@@ -77,15 +87,12 @@ func TestReportWriter_JSON(t *testing.T) {
 				},
 			}
 
-			err := report.Write(inputResults, report.Option{
-				Format: "json",
-				Output: &jsonWritten,
-			})
-			assert.NoError(t, err)
+			err := jw.Write(context.Background(), inputResults)
+			require.NoError(t, err)
 
-			var got report.Report
+			var got types.Report
 			err = json.Unmarshal(jsonWritten.Bytes(), &got)
-			assert.NoError(t, err, "invalid json written")
+			require.NoError(t, err, "invalid json written")
 
 			assert.Equal(t, tc.want, got, tc.name)
 		})
