@@ -1,14 +1,22 @@
 package report
 
 import (
+	"context"
 	"io"
-	"time"
+	"strings"
 
-	v1 "github.com/google/go-containerregistry/pkg/v1"
+	"github.com/hashicorp/go-multierror"
 	"golang.org/x/xerrors"
 
-	ftypes "github.com/aquasecurity/fanal/types"
-	dbTypes "github.com/aquasecurity/trivy-db/pkg/types"
+	cr "github.com/aquasecurity/trivy/pkg/compliance/report"
+	"github.com/aquasecurity/trivy/pkg/fanal/artifact"
+	"github.com/aquasecurity/trivy/pkg/flag"
+	"github.com/aquasecurity/trivy/pkg/log"
+	"github.com/aquasecurity/trivy/pkg/report/cyclonedx"
+	"github.com/aquasecurity/trivy/pkg/report/github"
+	"github.com/aquasecurity/trivy/pkg/report/predicate"
+	"github.com/aquasecurity/trivy/pkg/report/spdx"
+	"github.com/aquasecurity/trivy/pkg/report/table"
 	"github.com/aquasecurity/trivy/pkg/types"
 )
 
@@ -16,120 +24,102 @@ const (
 	SchemaVersion = 2
 )
 
-// Now returns the current time
-var Now = time.Now
-
-// Report represents a scan result
-type Report struct {
-	SchemaVersion int                 `json:",omitempty"`
-	ArtifactName  string              `json:",omitempty"`
-	ArtifactType  ftypes.ArtifactType `json:",omitempty"`
-	Metadata      Metadata            `json:",omitempty"`
-	Results       Results             `json:",omitempty"`
-}
-
-// Metadata represents a metadata of artifact
-type Metadata struct {
-	Size int64      `json:",omitempty"`
-	OS   *ftypes.OS `json:",omitempty"`
-
-	// Container image
-	ImageID     string        `json:",omitempty"`
-	DiffIDs     []string      `json:",omitempty"`
-	RepoTags    []string      `json:",omitempty"`
-	RepoDigests []string      `json:",omitempty"`
-	ImageConfig v1.ConfigFile `json:",omitempty"`
-}
-
-// Results to hold list of Result
-type Results []Result
-
-type ResultClass string
-
-const (
-	ClassOSPkg   = "os-pkgs"
-	ClassLangPkg = "lang-pkgs"
-	ClassConfig  = "config"
-)
-
-// Result holds a target and detected vulnerabilities
-type Result struct {
-	Target            string                           `json:"Target"`
-	Class             ResultClass                      `json:"Class,omitempty"`
-	Type              string                           `json:"Type,omitempty"`
-	Packages          []ftypes.Package                 `json:"Packages,omitempty"`
-	Vulnerabilities   []types.DetectedVulnerability    `json:"Vulnerabilities,omitempty"`
-	MisconfSummary    *MisconfSummary                  `json:"MisconfSummary,omitempty"`
-	Misconfigurations []types.DetectedMisconfiguration `json:"Misconfigurations,omitempty"`
-}
-
-type MisconfSummary struct {
-	Successes  int
-	Failures   int
-	Exceptions int
-}
-
-func (s MisconfSummary) Empty() bool {
-	return s.Successes == 0 && s.Failures == 0 && s.Exceptions == 0
-}
-
-// Failed returns whether the result includes any vulnerabilities or misconfigurations
-func (results Results) Failed() bool {
-	for _, r := range results {
-		if len(r.Vulnerabilities) > 0 {
-			return true
-		}
-		for _, m := range r.Misconfigurations {
-			if m.Status == types.StatusFailure {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-type Option struct {
-	Format         string
-	Output         io.Writer
-	Severities     []dbTypes.Severity
-	OutputTemplate string
-	Light          bool
-
-	// For misconfigurations
-	IncludeNonFailures bool
-	Trace              bool
-}
-
 // Write writes the result to output, format as passed in argument
-func Write(report Report, option Option) error {
+func Write(ctx context.Context, report types.Report, option flag.Options) (err error) {
+	output, cleanup, err := option.OutputWriter(ctx)
+	if err != nil {
+		return xerrors.Errorf("failed to create a file: %w", err)
+	}
+	defer func() {
+		if cerr := cleanup(); cerr != nil {
+			err = multierror.Append(err, cerr)
+		}
+	}()
+
+	// Compliance report
+	if option.Compliance.Spec.ID != "" {
+		return complianceWrite(ctx, report, option, output)
+	}
+
 	var writer Writer
 	switch option.Format {
-	case "table":
-		writer = &TableWriter{
-			Output:             option.Output,
-			Severities:         option.Severities,
-			Light:              option.Light,
-			IncludeNonFailures: option.IncludeNonFailures,
-			Trace:              option.Trace,
+	case types.FormatTable:
+		writer = &table.Writer{
+			Output:               output,
+			Severities:           option.Severities,
+			Tree:                 option.DependencyTree,
+			ShowSuppressed:       option.ShowSuppressed,
+			IncludeNonFailures:   option.IncludeNonFailures,
+			Trace:                option.Trace,
+			LicenseRiskThreshold: option.LicenseRiskThreshold,
+			IgnoredLicenses:      option.IgnoredLicenses,
 		}
-	case "json":
-		writer = &JSONWriter{Output: option.Output}
-	case "template":
-		var err error
-		if writer, err = NewTemplateWriter(option.Output, option.OutputTemplate); err != nil {
+	case types.FormatJSON:
+		writer = &JSONWriter{
+			Output:         output,
+			ListAllPkgs:    option.ListAllPkgs,
+			ShowSuppressed: option.ShowSuppressed,
+		}
+	case types.FormatGitHub:
+		writer = &github.Writer{
+			Output:  output,
+			Version: option.AppVersion,
+		}
+	case types.FormatCycloneDX:
+		// TODO: support xml format option with cyclonedx writer
+		writer = cyclonedx.NewWriter(output, option.AppVersion)
+	case types.FormatSPDX, types.FormatSPDXJSON:
+		writer = spdx.NewWriter(output, option.AppVersion, option.Format)
+	case types.FormatTemplate:
+		// We keep `sarif.tpl` template working for backward compatibility for a while.
+		if strings.HasPrefix(option.Template, "@") && strings.HasSuffix(option.Template, "sarif.tpl") {
+			log.Warn("Using `--template sarif.tpl` is deprecated. Please migrate to `--format sarif`. See https://github.com/aquasecurity/trivy/discussions/1571")
+			writer = &SarifWriter{
+				Output:  output,
+				Version: option.AppVersion,
+			}
+			break
+		}
+		if writer, err = NewTemplateWriter(output, option.Template, option.AppVersion); err != nil {
 			return xerrors.Errorf("failed to initialize template writer: %w", err)
 		}
+	case types.FormatSarif:
+		target := ""
+		if report.ArtifactType == artifact.TypeFilesystem {
+			target = option.Target
+		}
+		writer = &SarifWriter{
+			Output:  output,
+			Version: option.AppVersion,
+			Target:  target,
+		}
+	case types.FormatCosignVuln:
+		writer = predicate.NewVulnWriter(output, option.AppVersion)
 	default:
 		return xerrors.Errorf("unknown format: %v", option.Format)
 	}
 
-	if err := writer.Write(report); err != nil {
+	if err = writer.Write(ctx, report); err != nil {
 		return xerrors.Errorf("failed to write results: %w", err)
 	}
+
 	return nil
+}
+
+func complianceWrite(ctx context.Context, report types.Report, opt flag.Options, output io.Writer) error {
+	complianceReport, err := cr.BuildComplianceReport([]types.Results{report.Results}, opt.Compliance)
+	if err != nil {
+		return xerrors.Errorf("compliance report build error: %w", err)
+	}
+	return cr.Write(ctx, complianceReport, cr.Option{
+		Format:     opt.Format,
+		Report:     opt.ReportFormat,
+		Output:     output,
+		Severities: opt.Severities,
+	})
 }
 
 // Writer defines the result write operation
 type Writer interface {
-	Write(Report) error
+	Write(context.Context, types.Report) error
 }
