@@ -5,19 +5,18 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/aquasecurity/trivy/pkg/report"
 	"github.com/aquasecurity/trivy/pkg/types"
 )
 
 func TestResults_Failed(t *testing.T) {
 	tests := []struct {
 		name    string
-		results report.Results
+		results types.Results
 		want    bool
 	}{
 		{
 			name: "no vulnerabilities and misconfigurations",
-			results: report.Results{
+			results: types.Results{
 				{
 					Target: "test",
 					Type:   "test",
@@ -27,7 +26,7 @@ func TestResults_Failed(t *testing.T) {
 		},
 		{
 			name: "vulnerabilities found",
-			results: report.Results{
+			results: types.Results{
 				{
 					Target: "test",
 					Type:   "test",
@@ -43,7 +42,7 @@ func TestResults_Failed(t *testing.T) {
 		},
 		{
 			name: "failed misconfigurations",
-			results: report.Results{
+			results: types.Results{
 				{
 					Target: "test",
 					Type:   "test",
@@ -51,7 +50,7 @@ func TestResults_Failed(t *testing.T) {
 						{
 							Type:   "Docker Security Check",
 							ID:     "ID-001",
-							Status: types.StatusFailure,
+							Status: types.MisconfStatusFailure,
 						},
 					},
 				},
@@ -60,7 +59,7 @@ func TestResults_Failed(t *testing.T) {
 		},
 		{
 			name: "passed misconfigurations",
-			results: report.Results{
+			results: types.Results{
 				{
 					Target: "test",
 					Type:   "test",
@@ -68,7 +67,7 @@ func TestResults_Failed(t *testing.T) {
 						{
 							Type:   "Docker Security Check",
 							ID:     "ID-001",
-							Status: types.StatusPassed,
+							Status: types.MisconfStatusPassed,
 						},
 					},
 				},
