@@ -2,163 +2,120 @@ package server
 
 import (
 	"context"
+	"encoding/json"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"path"
-	"path/filepath"
 	"sync"
 	"testing"
 	"time"
 
+	"github.com/google/go-containerregistry/pkg/v1/types"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"golang.org/x/xerrors"
 
-	"github.com/aquasecurity/fanal/cache"
-	"github.com/aquasecurity/trivy-db/pkg/db"
-	dbFile "github.com/aquasecurity/trivy/pkg/db"
+	"github.com/aquasecurity/trivy-db/pkg/metadata"
+	"github.com/aquasecurity/trivy/internal/dbtest"
+	"github.com/aquasecurity/trivy/pkg/cache"
+	"github.com/aquasecurity/trivy/pkg/clock"
+	"github.com/aquasecurity/trivy/pkg/db"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
+	"github.com/aquasecurity/trivy/pkg/policy"
+	"github.com/aquasecurity/trivy/pkg/version"
 	rpcCache "github.com/aquasecurity/trivy/rpc/cache"
 )
 
 func Test_dbWorker_update(t *testing.T) {
-	timeNextUpdate := time.Date(3000, 1, 1, 0, 0, 0, 0, time.UTC)
-	timeUpdateAt := time.Date(3000, 1, 1, 0, 0, 0, 0, time.UTC)
-
-	type needsUpdateInput struct {
-		appVersion string
-		skip       bool
-	}
-	type needsUpdateOutput struct {
-		needsUpdate bool
-		err         error
-	}
-	type needsUpdate struct {
-		input  needsUpdateInput
-		output needsUpdateOutput
+	cachedMetadata := metadata.Metadata{
+		Version:      db.SchemaVersion,
+		NextUpdate:   time.Date(2020, 10, 2, 0, 0, 0, 0, time.UTC),
+		UpdatedAt:    time.Date(2020, 10, 1, 0, 0, 0, 0, time.UTC),
+		DownloadedAt: time.Date(2020, 10, 1, 1, 0, 0, 0, time.UTC),
 	}
 
-	type download struct {
-		call bool
-		err  error
-	}
-
-	type args struct {
-		appVersion string
-	}
 	tests := []struct {
-		name        string
-		needsUpdate needsUpdate
-		download    download
-		args        args
-		want        db.Metadata
-		wantErr     string
+		name           string
+		now            time.Time
+		skipUpdate     bool
+		layerMediaType types.MediaType
+		want           metadata.Metadata
+		wantErr        string
 	}{
 		{
-			name: "happy path",
-			needsUpdate: needsUpdate{
-				input:  needsUpdateInput{appVersion: "1", skip: false},
-				output: needsUpdateOutput{needsUpdate: true},
-			},
-			download: download{
-				call: true,
-			},
-			args: args{appVersion: "1"},
-			want: db.Metadata{
-				Version:    1,
-				Type:       db.TypeFull,
-				NextUpdate: timeNextUpdate,
-				UpdatedAt:  timeUpdateAt,
+			name:       "update needed",
+			now:        time.Date(2021, 10, 1, 0, 0, 0, 0, time.UTC),
+			skipUpdate: false,
+			want: metadata.Metadata{
+				Version:      db.SchemaVersion,
+				NextUpdate:   time.Date(3000, 1, 1, 0, 0, 0, 0, time.UTC),
+				UpdatedAt:    time.Date(3000, 1, 1, 0, 0, 0, 0, time.UTC),
+				DownloadedAt: time.Date(2021, 10, 1, 0, 0, 0, 0, time.UTC),
 			},
 		},
 		{
-			name: "not update",
-			needsUpdate: needsUpdate{
-				input:  needsUpdateInput{appVersion: "1", skip: false},
-				output: needsUpdateOutput{needsUpdate: false},
-			},
-			args: args{appVersion: "1"},
+			name:       "not update needed",
+			now:        time.Date(2019, 10, 1, 0, 0, 0, 0, time.UTC),
+			skipUpdate: false,
+			want:       cachedMetadata,
 		},
 		{
-			name: "NeedsUpdate returns an error",
-			needsUpdate: needsUpdate{
-				input:  needsUpdateInput{appVersion: "1", skip: false},
-				output: needsUpdateOutput{err: xerrors.New("fail")},
-			},
-			args:    args{appVersion: "1"},
-			wantErr: "failed to check if db needs an update",
+			name:       "skip update",
+			now:        time.Date(2021, 10, 1, 0, 0, 0, 0, time.UTC),
+			skipUpdate: true,
+			want:       cachedMetadata,
 		},
 		{
-			name: "Download returns an error",
-			needsUpdate: needsUpdate{
-				input:  needsUpdateInput{appVersion: "1", skip: false},
-				output: needsUpdateOutput{needsUpdate: true},
-			},
-			download: download{
-				call: true,
-				err:  xerrors.New("fail"),
-			},
-			args:    args{appVersion: "1"},
-			wantErr: "failed DB hot update",
+			name:           "Download returns an error",
+			now:            time.Date(2021, 10, 1, 0, 0, 0, 0, time.UTC),
+			skipUpdate:     false,
+			layerMediaType: types.MediaType("unknown"),
+			wantErr:        "failed DB hot update",
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			cacheDir, err := os.MkdirTemp("", "server-test")
-			require.NoError(t, err, tt.name)
+			dbDir := db.Dir(t.TempDir())
 
-			require.NoError(t, db.Init(cacheDir), tt.name)
+			// Initialize the cache
+			meta := metadata.NewClient(dbDir)
+			err := meta.Update(cachedMetadata)
+			require.NoError(t, err)
 
-			mockDBClient := new(dbFile.MockOperation)
-			mockDBClient.On("NeedsUpdate",
-				tt.needsUpdate.input.appVersion, false, tt.needsUpdate.input.skip).Return(
-				tt.needsUpdate.output.needsUpdate, tt.needsUpdate.output.err)
-			mockDBClient.On("UpdateMetadata", mock.Anything).Return(nil)
+			err = db.Init(dbDir)
+			require.NoError(t, err)
 
-			if tt.download.call {
-				mockDBClient.On("Download", mock.Anything, mock.Anything, false).Run(
-					func(args mock.Arguments) {
-						// fake download: copy testdata/new.db to tmpDir/db/trivy.db
-						content, err := os.ReadFile("testdata/new.db")
-						require.NoError(t, err, tt.name)
+			defer func() { _ = db.Close() }()
 
-						tmpDir := args.String(1)
-						dbPath := db.Path(tmpDir)
-						require.NoError(t, os.MkdirAll(filepath.Dir(dbPath), 0777), tt.name)
-						err = os.WriteFile(dbPath, content, 0444)
-						require.NoError(t, err, tt.name)
-					}).Return(tt.download.err)
-			}
+			// Set a fake time
+			ctx := clock.With(context.Background(), tt.now)
 
-			w := newDBWorker(mockDBClient)
+			// Set a fake DB
+			dbPath := dbtest.ArchiveDir(t, "testdata/newdb")
+			art := dbtest.NewFakeDB(t, dbPath, dbtest.FakeDBOptions{
+				MediaType: tt.layerMediaType,
+			})
+			client := db.NewClient(dbDir, true, db.WithOCIArtifact(art))
+			w := newDBWorker(client)
 
 			var dbUpdateWg, requestWg sync.WaitGroup
-			err = w.update(context.Background(), tt.args.appVersion, cacheDir,
-				&dbUpdateWg, &requestWg)
+			err = w.update(ctx, "1.2.3", dbDir,
+				tt.skipUpdate, &dbUpdateWg, &requestWg, ftypes.RegistryOptions{})
 			if tt.wantErr != "" {
-				require.NotNil(t, err, tt.name)
+				require.Error(t, err, tt.name)
 				assert.Contains(t, err.Error(), tt.wantErr, tt.name)
 				return
-			} else {
-				assert.NoError(t, err, tt.name)
 			}
+			require.NoError(t, err, tt.name)
 
-			if !tt.download.call {
-				return
-			}
-
-			dbc := db.Config{}
-			got, err := dbc.GetMetadata()
-			assert.NoError(t, err, tt.name)
+			mc := metadata.NewClient(dbDir)
+			got, err := mc.Get()
+			require.NoError(t, err, tt.name)
 			assert.Equal(t, tt.want, got, tt.name)
-
-			mockDBClient.AssertExpectations(t)
 		})
 	}
 }
 
-func Test_newServeMux(t *testing.T) {
+func TestServer_newServeMux(t *testing.T) {
 	type args struct {
 		token       string
 		tokenHeader string
@@ -223,28 +180,64 @@ func Test_newServeMux(t *testing.T) {
 
 			c, err := cache.NewFSCache(t.TempDir())
 			require.NoError(t, err)
+			defer func() { _ = c.Close() }()
 
-			ts := httptest.NewServer(newServeMux(
-				c, dbUpdateWg, requestWg, tt.args.token, tt.args.tokenHeader),
-			)
+			s := NewServer("", "", "", tt.args.token, tt.args.tokenHeader, "", nil, ftypes.RegistryOptions{})
+			ts := httptest.NewServer(s.NewServeMux(context.Background(), c, dbUpdateWg, requestWg))
 			defer ts.Close()
 
 			var resp *http.Response
 			url := ts.URL + tt.path
 			if tt.header == nil {
 				resp, err = http.Get(url)
+				require.NoError(t, err)
+				defer resp.Body.Close()
 			} else {
-				req, err := http.NewRequest(http.MethodPost, url, nil)
+				req, err := http.NewRequest(http.MethodPost, url, http.NoBody)
 				require.NoError(t, err)
 
 				req.Header = tt.header
 				client := new(http.Client)
 				resp, err = client.Do(req)
+				require.NoError(t, err)
+				defer resp.Body.Close()
 			}
-
-			require.NoError(t, err)
 			assert.Equal(t, tt.want, resp.StatusCode)
-			defer resp.Body.Close()
 		})
 	}
+}
+
+func Test_VersionEndpoint(t *testing.T) {
+	dbUpdateWg, requestWg := &sync.WaitGroup{}, &sync.WaitGroup{}
+	c, err := cache.NewFSCache(t.TempDir())
+	require.NoError(t, err)
+	defer func() { _ = c.Close() }()
+
+	s := NewServer("", "", "testdata/testcache", "", "", "", nil, ftypes.RegistryOptions{})
+	ts := httptest.NewServer(s.NewServeMux(context.Background(), c, dbUpdateWg, requestWg))
+	defer ts.Close()
+
+	resp, err := http.Get(ts.URL + "/version")
+	require.NoError(t, err)
+	defer resp.Body.Close()
+
+	assert.Equal(t, http.StatusOK, resp.StatusCode)
+
+	var versionInfo version.VersionInfo
+	require.NoError(t, json.NewDecoder(resp.Body).Decode(&versionInfo))
+
+	expected := version.VersionInfo{
+		Version: "dev",
+		VulnerabilityDB: &metadata.Metadata{
+			Version:      2,
+			NextUpdate:   time.Date(2023, 7, 20, 18, 11, 37, 696263532, time.UTC),
+			UpdatedAt:    time.Date(2023, 7, 20, 12, 11, 37, 696263932, time.UTC),
+			DownloadedAt: time.Date(2023, 7, 25, 7, 1, 41, 239158000, time.UTC),
+		},
+		CheckBundle: &policy.Metadata{
+			Digest:       "sha256:829832357626da2677955e3b427191212978ba20012b6eaa03229ca28569ae43",
+			DownloadedAt: time.Date(2023, 7, 23, 16, 40, 33, 122462000, time.UTC),
+		},
+	}
+	assert.Equal(t, expected, versionInfo)
 }
