@@ -3,8 +3,9 @@
 package local
 
 import (
-	types "github.com/aquasecurity/fanal/types"
 	mock "github.com/stretchr/testify/mock"
+
+	types "github.com/aquasecurity/trivy/pkg/fanal/types"
 )
 
 // MockApplier is an autogenerated mock type for the Applier type
