@@ -1,20 +1,104 @@
 package local
 
 import (
+	"context"
 	"errors"
 	"testing"
+	"time"
 
+	"github.com/samber/lo"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/aquasecurity/fanal/analyzer"
-	ftypes "github.com/aquasecurity/fanal/types"
 	"github.com/aquasecurity/trivy-db/pkg/db"
-	"github.com/aquasecurity/trivy-db/pkg/vulnsrc/vulnerability"
-	"github.com/aquasecurity/trivy/pkg/dbtest"
-	ospkgDetector "github.com/aquasecurity/trivy/pkg/detector/ospkg"
-	"github.com/aquasecurity/trivy/pkg/report"
+	dbTypes "github.com/aquasecurity/trivy-db/pkg/types"
+	"github.com/aquasecurity/trivy/internal/dbtest"
+	"github.com/aquasecurity/trivy/pkg/fanal/analyzer"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
+	"github.com/aquasecurity/trivy/pkg/scanner/langpkg"
+	"github.com/aquasecurity/trivy/pkg/scanner/ospkg"
 	"github.com/aquasecurity/trivy/pkg/types"
+	"github.com/aquasecurity/trivy/pkg/vulnerability"
+)
+
+var (
+	muslPkg = ftypes.Package{
+		Name:       "musl",
+		Version:    "1.2.3",
+		SrcName:    "musl",
+		SrcVersion: "1.2.3",
+		Licenses:   []string{"MIT"},
+		Layer: ftypes.Layer{
+			DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
+		},
+	}
+	railsPkg = ftypes.Package{
+		Name:    "rails",
+		Version: "4.0.2",
+		Layer: ftypes.Layer{
+			DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+		},
+	}
+	innocentPkg = ftypes.Package{
+		Name:    "innocent",
+		Version: "1.2.3",
+		Layer: ftypes.Layer{
+			DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+		},
+	}
+	uuidPkg = ftypes.Package{
+		Name:     "github.com/google/uuid",
+		Version:  "1.6.0",
+		FilePath: "",
+		Layer: ftypes.Layer{
+			DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+		},
+		Licenses: []string{"LGPL"},
+	}
+	urllib3Pkg = ftypes.Package{
+		Name:     "urllib3",
+		Version:  "3.2.1",
+		FilePath: "/usr/lib/python/site-packages/urllib3-3.2.1/METADATA",
+		Layer: ftypes.Layer{
+			DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+		},
+		Licenses: []string{"MIT"},
+	}
+	python39min = ftypes.Package{
+		Name:     "python3.9-minimal",
+		Version:  "3.9.1",
+		FilePath: "/usr/lib/python/site-packages/python3.9-minimal/METADATA",
+		Layer: ftypes.Layer{
+			DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+		},
+		Licenses: []string{"text://Redistribution and use in source and binary forms, with or without"},
+	}
+	menuinstPkg = ftypes.Package{
+		Name:     "menuinst",
+		Version:  "2.0.2",
+		FilePath: "opt/conda/lib/python3.11/site-packages/menuinst-2.0.2.dist-info/METADATA",
+		Layer: ftypes.Layer{
+			DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+		},
+		Licenses: []string{"text://(c) 2016 Continuum Analytics, Inc. / http://continuum.io All Rights Reserved"},
+	}
+
+	laravelPkg = ftypes.Package{
+		Name:         "laravel/framework",
+		Version:      "6.0.0",
+		Relationship: ftypes.RelationshipDirect,
+		Layer: ftypes.Layer{
+			DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+		},
+	}
+	guzzlePkg = ftypes.Package{
+		Name:         "guzzlehttp/guzzle",
+		Version:      "7.9.2",
+		Relationship: ftypes.RelationshipIndirect,
+		Layer: ftypes.Layer{
+			DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+		},
+	}
 )
 
 func TestScanner_Scan(t *testing.T) {
@@ -24,14 +108,13 @@ func TestScanner_Scan(t *testing.T) {
 		options  types.ScanOptions
 	}
 	tests := []struct {
-		name                    string
-		args                    args
-		fixtures                []string
-		applyLayersExpectation  ApplierApplyLayersExpectation
-		ospkgDetectExpectations []OspkgDetectorDetectExpectation
-		wantResults             report.Results
-		wantOS                  *ftypes.OS
-		wantErr                 string
+		name                   string
+		args                   args
+		fixtures               []string
+		applyLayersExpectation ApplierApplyLayersExpectation
+		wantResults            types.Results
+		wantOS                 ftypes.OS
+		wantErr                string
 	}{
 		{
 			name: "happy path",
@@ -39,8 +122,12 @@ func TestScanner_Scan(t *testing.T) {
 				target:   "alpine:latest",
 				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
 				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeOS, types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
+					PkgTypes: []string{
+						types.PkgTypeOS,
+						types.PkgTypeLibrary,
+					},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
@@ -50,117 +137,106 @@ func TestScanner_Scan(t *testing.T) {
 				},
 				Returns: ApplierApplyLayersReturns{
 					Detail: ftypes.ArtifactDetail{
-						OS: &ftypes.OS{
-							Family: "alpine",
+						OS: ftypes.OS{
+							Family: ftypes.Alpine,
 							Name:   "3.11",
 						},
 						Packages: []ftypes.Package{
-							{
-								Name:    "musl",
-								Version: "1.2.3",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
+							muslPkg,
 						},
 						Applications: []ftypes.Application{
 							{
 								Type:     ftypes.Bundler,
 								FilePath: "/app/Gemfile.lock",
-								Libraries: []ftypes.Package{
-									{
-										Name:    "rails",
-										Version: "4.0.2",
-										Layer: ftypes.Layer{
-											DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
-										},
-									},
+								Packages: []ftypes.Package{
+									railsPkg,
 								},
 							},
 						},
 					},
 				},
 			},
-			ospkgDetectExpectations: []OspkgDetectorDetectExpectation{
-				{
-					Args: OspkgDetectorDetectArgs{
-						OsFamily: "alpine",
-						OsName:   "3.11",
-						Pkgs: []ftypes.Package{
-							{
-								Name:    "musl",
-								Version: "1.2.3",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-						},
-					},
-					Returns: OspkgDetectorDetectReturns{
-						DetectedVulns: []types.DetectedVulnerability{
-							{
-								VulnerabilityID:  "CVE-2020-9999",
-								PkgName:          "musl",
-								InstalledVersion: "1.2.3",
-								FixedVersion:     "1.2.4",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-						},
-						Eosl: true,
-					},
-				},
-			},
-			wantResults: report.Results{
+			wantResults: types.Results{
 				{
 					Target: "alpine:latest (alpine 3.11)",
+					Class:  types.ClassOSPkg,
+					Type:   ftypes.Alpine,
+					Packages: ftypes.Packages{
+						muslPkg,
+					},
 					Vulnerabilities: []types.DetectedVulnerability{
 						{
 							VulnerabilityID:  "CVE-2020-9999",
-							PkgName:          "musl",
-							InstalledVersion: "1.2.3",
+							PkgName:          muslPkg.Name,
+							InstalledVersion: muslPkg.Version,
 							FixedVersion:     "1.2.4",
+							Status:           dbTypes.StatusFixed,
 							Layer: ftypes.Layer{
 								DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
 							},
-						},
-					},
-					Class: report.ClassOSPkg,
-					Type:  vulnerability.Alpine,
-				},
-				{
-					Target: "/app/Gemfile.lock",
-					Vulnerabilities: []types.DetectedVulnerability{
-						{
-							VulnerabilityID:  "CVE-2014-0081",
-							PkgName:          "rails",
-							InstalledVersion: "4.0.2",
-							FixedVersion:     "4.0.3, 3.2.17",
-							Layer: ftypes.Layer{
-								DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+							PrimaryURL: "https://avd.aquasec.com/nvd/cve-2020-9999",
+							Vulnerability: dbTypes.Vulnerability{
+								Title:       "dos",
+								Description: "dos vulnerability",
+								Severity:    "HIGH",
 							},
 						},
 					},
-					Class: report.ClassLangPkg,
-					Type:  ftypes.Bundler,
+				},
+				{
+					Target: "/app/Gemfile.lock",
+					Class:  types.ClassLangPkg,
+					Type:   ftypes.Bundler,
+					Packages: ftypes.Packages{
+						railsPkg,
+					},
+					Vulnerabilities: []types.DetectedVulnerability{
+						{
+							VulnerabilityID:  "CVE-2014-0081",
+							PkgName:          railsPkg.Name,
+							InstalledVersion: railsPkg.Version,
+							FixedVersion:     "4.0.3, 3.2.17",
+							Status:           dbTypes.StatusFixed,
+							Layer: ftypes.Layer{
+								DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+							},
+							PrimaryURL: "https://avd.aquasec.com/nvd/cve-2014-0081",
+							Vulnerability: dbTypes.Vulnerability{
+								Title:       "xss",
+								Description: "xss vulnerability",
+								Severity:    "MEDIUM",
+								References: []string{
+									"http://example.com",
+								},
+								LastModifiedDate: lo.ToPtr(time.Date(2020, 2, 1, 1, 1, 0, 0, time.UTC)),
+								PublishedDate:    lo.ToPtr(time.Date(2020, 1, 1, 1, 1, 0, 0, time.UTC)),
+							},
+						},
+					},
 				},
 			},
-			wantOS: &ftypes.OS{
+			wantOS: ftypes.OS{
 				Family: "alpine",
 				Name:   "3.11",
 				Eosl:   true,
 			},
 		},
 		{
-			name: "happy path with list all packages",
+			name: "happy path with OS rewriting",
 			args: args{
 				target:   "alpine:latest",
 				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
 				options: types.ScanOptions{
-					VulnType:        []string{types.VulnTypeOS, types.VulnTypeLibrary},
-					SecurityChecks:  []string{types.SecurityCheckVulnerability},
-					ListAllPackages: true,
+					PkgTypes: []string{
+						types.PkgTypeOS,
+						types.PkgTypeLibrary,
+					},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
+					Distro: ftypes.OS{
+						Family: "alpine",
+						Name:   "3.11",
+					},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
@@ -170,144 +246,164 @@ func TestScanner_Scan(t *testing.T) {
 				},
 				Returns: ApplierApplyLayersReturns{
 					Detail: ftypes.ArtifactDetail{
-						OS: &ftypes.OS{
-							Family: "alpine",
-							Name:   "3.11",
+						OS: ftypes.OS{
+							Family: ftypes.Alpine,
+							Name:   "3.10",
 						},
 						Packages: []ftypes.Package{
-							{
-								Name:    "musl",
-								Version: "1.2.3",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-							{
-								Name:    "ausl",
-								Version: "1.2.3",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:bbf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-						},
-						Applications: []ftypes.Application{
-							{
-								Type:     "bundler",
-								FilePath: "/app/Gemfile.lock",
-								Libraries: []ftypes.Package{
-									{
-										Name:    "rails",
-										Version: "4.0.2",
-										Layer: ftypes.Layer{
-											DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
-										},
-									},
-								},
-							},
+							muslPkg,
 						},
 					},
 				},
 			},
-			ospkgDetectExpectations: []OspkgDetectorDetectExpectation{
-				{
-					Args: OspkgDetectorDetectArgs{
-						OsFamily: "alpine",
-						OsName:   "3.11",
-						Pkgs: []ftypes.Package{
-							{
-								Name:    "musl",
-								Version: "1.2.3",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-							{
-								Name:    "ausl",
-								Version: "1.2.3",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:bbf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-						},
-					},
-					Returns: OspkgDetectorDetectReturns{
-						DetectedVulns: []types.DetectedVulnerability{
-							{
-								VulnerabilityID:  "CVE-2020-9999",
-								PkgName:          "musl",
-								InstalledVersion: "1.2.3",
-								FixedVersion:     "1.2.4",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-						},
-						Eosl: false,
-					},
-				},
-			},
-			wantResults: report.Results{
+			wantResults: types.Results{
 				{
 					Target: "alpine:latest (alpine 3.11)",
-					Packages: []ftypes.Package{
-						{
-							Name:    "ausl",
-							Version: "1.2.3",
-							Layer: ftypes.Layer{
-								DiffID: "sha256:bbf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-							},
-						},
-						{
-							Name:    "musl",
-							Version: "1.2.3",
-							Layer: ftypes.Layer{
-								DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-							},
-						},
+					Class:  types.ClassOSPkg,
+					Type:   ftypes.Alpine,
+					Packages: ftypes.Packages{
+						muslPkg,
 					},
 					Vulnerabilities: []types.DetectedVulnerability{
 						{
 							VulnerabilityID:  "CVE-2020-9999",
-							PkgName:          "musl",
-							InstalledVersion: "1.2.3",
+							PkgName:          muslPkg.Name,
+							InstalledVersion: muslPkg.Version,
 							FixedVersion:     "1.2.4",
+							Status:           dbTypes.StatusFixed,
 							Layer: ftypes.Layer{
 								DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
 							},
-						},
-					},
-					Class: report.ClassOSPkg,
-					Type:  vulnerability.Alpine,
-				},
-				{
-					Target: "/app/Gemfile.lock",
-					Packages: []ftypes.Package{
-						{
-							Name:    "rails",
-							Version: "4.0.2",
-							Layer: ftypes.Layer{
-								DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+							PrimaryURL: "https://avd.aquasec.com/nvd/cve-2020-9999",
+							Vulnerability: dbTypes.Vulnerability{
+								Title:       "dos",
+								Description: "dos vulnerability",
+								Severity:    "HIGH",
 							},
 						},
 					},
-					Vulnerabilities: []types.DetectedVulnerability{
-						{
-							VulnerabilityID:  "CVE-2014-0081",
-							PkgName:          "rails",
-							InstalledVersion: "4.0.2",
-							FixedVersion:     "4.0.3, 3.2.17",
-							Layer: ftypes.Layer{
-								DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
-							},
-						},
-					},
-					Class: report.ClassLangPkg,
-					Type:  ftypes.Bundler,
 				},
 			},
-			wantOS: &ftypes.OS{
+			wantOS: ftypes.OS{
 				Family: "alpine",
 				Name:   "3.11",
+				Eosl:   true,
+			},
+		},
+		{
+			name: "happy path license scanner",
+			args: args{
+				target:   "alpine:latest",
+				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
+				options: types.ScanOptions{
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.LicenseScanner},
+				},
+			},
+			fixtures: []string{"testdata/fixtures/happy.yaml"},
+			applyLayersExpectation: ApplierApplyLayersExpectation{
+				Args: ApplierApplyLayersArgs{
+					BlobIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
+				},
+				Returns: ApplierApplyLayersReturns{
+					Detail: ftypes.ArtifactDetail{
+						OS: ftypes.OS{
+							Family: ftypes.Alpine,
+							Name:   "3.11",
+						},
+						Packages: []ftypes.Package{
+							muslPkg,
+							python39min,
+						},
+						Applications: []ftypes.Application{
+							{
+								Type:     ftypes.GoModule,
+								FilePath: "/app/go.mod",
+								Packages: []ftypes.Package{
+									uuidPkg,
+								},
+							},
+							{
+								Type:     ftypes.PythonPkg,
+								FilePath: "",
+								Packages: []ftypes.Package{
+									urllib3Pkg,
+									menuinstPkg,
+								},
+							},
+						},
+					},
+				},
+			},
+			wantResults: types.Results{
+				{
+					Target: "OS Packages",
+					Class:  types.ClassLicense,
+					Licenses: []types.DetectedLicense{
+						{
+							Severity:   "UNKNOWN",
+							Category:   "unknown",
+							PkgName:    muslPkg.Name,
+							Name:       "MIT",
+							Confidence: 1,
+						},
+						{
+							Severity:   "UNKNOWN",
+							Category:   "unknown",
+							PkgName:    python39min.Name,
+							Name:       "CUSTOM License: Redistribution and use...",
+							Text:       "Redistribution and use in source and binary forms, with or without",
+							Confidence: 1,
+						},
+					},
+				},
+				{
+					Target: "/app/go.mod",
+					Class:  types.ClassLicense,
+					Licenses: []types.DetectedLicense{
+						{
+							Severity:   "UNKNOWN",
+							Category:   "unknown",
+							PkgName:    uuidPkg.Name,
+							FilePath:   "/app/go.mod",
+							Name:       "LGPL",
+							Confidence: 1,
+							Link:       "",
+						},
+					},
+				},
+				{
+					Target: "Python",
+					Class:  types.ClassLicense,
+					Licenses: []types.DetectedLicense{
+						{
+							Severity:   "UNKNOWN",
+							Category:   "unknown",
+							PkgName:    urllib3Pkg.Name,
+							FilePath:   "/usr/lib/python/site-packages/urllib3-3.2.1/METADATA",
+							Name:       "MIT",
+							Confidence: 1,
+						},
+						{
+							Severity:   "UNKNOWN",
+							Category:   "unknown",
+							PkgName:    menuinstPkg.Name,
+							FilePath:   "opt/conda/lib/python3.11/site-packages/menuinst-2.0.2.dist-info/METADATA",
+							Name:       "CUSTOM License: (c) 2016 Continuum...",
+							Text:       "(c) 2016 Continuum Analytics, Inc. / http://continuum.io All Rights Reserved",
+							Confidence: 1,
+						},
+					},
+				},
+				{
+					Target: "Loose File License(s)",
+					Class:  types.ClassLicenseFile,
+				},
+			},
+			wantOS: ftypes.OS{
+				Family: "alpine",
+				Name:   "3.11",
+				Eosl:   false,
 			},
 		},
 		{
@@ -316,8 +412,12 @@ func TestScanner_Scan(t *testing.T) {
 				target:   "alpine:latest",
 				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
 				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeOS, types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
+					PkgTypes: []string{
+						types.PkgTypeOS,
+						types.PkgTypeLibrary,
+					},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
@@ -327,18 +427,105 @@ func TestScanner_Scan(t *testing.T) {
 				},
 				Returns: ApplierApplyLayersReturns{
 					Detail: ftypes.ArtifactDetail{
-						OS: &ftypes.OS{},
+						OS: ftypes.OS{},
+						Applications: []ftypes.Application{
+							{
+								Type:     ftypes.Bundler,
+								FilePath: "/app1/Gemfile.lock",
+								Packages: []ftypes.Package{
+									innocentPkg, // no vulnerability
+								},
+							},
+							{
+								Type:     ftypes.Bundler,
+								FilePath: "/app2/Gemfile.lock",
+								Packages: []ftypes.Package{
+									railsPkg, // one vulnerability
+								},
+							},
+						},
+					},
+				},
+			},
+			wantResults: types.Results{
+				{
+					Target: "/app1/Gemfile.lock",
+					Class:  types.ClassLangPkg,
+					Type:   ftypes.Bundler,
+					Packages: ftypes.Packages{
+						innocentPkg,
+					},
+				},
+				{
+					Target: "/app2/Gemfile.lock",
+					Class:  types.ClassLangPkg,
+					Type:   ftypes.Bundler,
+					Packages: ftypes.Packages{
+						railsPkg,
+					},
+					Vulnerabilities: []types.DetectedVulnerability{
+						{
+							VulnerabilityID:  "CVE-2014-0081",
+							PkgName:          railsPkg.Name,
+							InstalledVersion: railsPkg.Version,
+							FixedVersion:     "4.0.3, 3.2.17",
+							Status:           dbTypes.StatusFixed,
+							Layer: ftypes.Layer{
+								DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+							},
+							PrimaryURL: "https://avd.aquasec.com/nvd/cve-2014-0081",
+							Vulnerability: dbTypes.Vulnerability{
+								Title:       "xss",
+								Description: "xss vulnerability",
+								Severity:    "MEDIUM",
+								References: []string{
+									"http://example.com",
+								},
+								LastModifiedDate: lo.ToPtr(time.Date(2020, 2, 1, 1, 1, 0, 0, time.UTC)),
+								PublishedDate:    lo.ToPtr(time.Date(2020, 1, 1, 1, 1, 0, 0, time.UTC)),
+							},
+						},
+					},
+				},
+			},
+			wantOS: ftypes.OS{},
+		},
+		{
+			name: "happy path. Empty filePaths (e.g. Scanned SBOM)",
+			args: args{
+				target:   "./result.cdx",
+				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
+				options: types.ScanOptions{
+					PkgTypes:         []string{types.PkgTypeLibrary},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
+				},
+			},
+			fixtures: []string{"testdata/fixtures/happy.yaml"},
+			applyLayersExpectation: ApplierApplyLayersExpectation{
+				Args: ApplierApplyLayersArgs{
+					BlobIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
+				},
+				Returns: ApplierApplyLayersReturns{
+					Detail: ftypes.ArtifactDetail{
 						Applications: []ftypes.Application{
 							{
 								Type:     "bundler",
-								FilePath: "/app/Gemfile.lock",
-								Libraries: []ftypes.Package{
+								FilePath: "",
+								Packages: []ftypes.Package{
 									{
 										Name:    "rails",
 										Version: "4.0.2",
-										Layer: ftypes.Layer{
-											DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
-										},
+									},
+								},
+							},
+							{
+								Type:     "composer",
+								FilePath: "",
+								Packages: []ftypes.Package{
+									{
+										Name:    "laravel/framework",
+										Version: "6.0.0",
 									},
 								},
 							},
@@ -346,25 +533,59 @@ func TestScanner_Scan(t *testing.T) {
 					},
 				},
 			},
-			wantResults: report.Results{
+			wantResults: types.Results{
 				{
-					Target: "/app/Gemfile.lock",
+					Target: "",
+					Class:  types.ClassLangPkg,
+					Type:   ftypes.Bundler,
+					Packages: []ftypes.Package{
+						{
+							Name:    "rails",
+							Version: "4.0.2",
+						},
+					},
 					Vulnerabilities: []types.DetectedVulnerability{
 						{
 							VulnerabilityID:  "CVE-2014-0081",
 							PkgName:          "rails",
 							InstalledVersion: "4.0.2",
 							FixedVersion:     "4.0.3, 3.2.17",
-							Layer: ftypes.Layer{
-								DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
+							Status:           dbTypes.StatusFixed,
+							PrimaryURL:       "https://avd.aquasec.com/nvd/cve-2014-0081",
+							Vulnerability: dbTypes.Vulnerability{
+								Title:       "xss",
+								Description: "xss vulnerability",
+								Severity:    "MEDIUM",
+								References: []string{
+									"http://example.com",
+								},
+								LastModifiedDate: lo.ToPtr(time.Date(2020, 2, 1, 1, 1, 0, 0, time.UTC)),
+								PublishedDate:    lo.ToPtr(time.Date(2020, 1, 1, 1, 1, 0, 0, time.UTC)),
 							},
 						},
 					},
-					Class: report.ClassLangPkg,
-					Type:  "bundler",
+				},
+				{
+					Target: "",
+					Class:  types.ClassLangPkg,
+					Type:   ftypes.Composer,
+					Packages: []ftypes.Package{
+						{
+							Name:    "laravel/framework",
+							Version: "6.0.0",
+						},
+					},
+					Vulnerabilities: []types.DetectedVulnerability{
+						{
+							VulnerabilityID:  "CVE-2021-21263",
+							PkgName:          "laravel/framework",
+							InstalledVersion: "6.0.0",
+							FixedVersion:     "8.22.1, 7.30.3, 6.20.12",
+							Status:           dbTypes.StatusFixed,
+						},
+					},
 				},
 			},
-			wantOS: &ftypes.OS{},
 		},
 		{
 			name: "happy path with no package",
@@ -372,8 +593,12 @@ func TestScanner_Scan(t *testing.T) {
 				target:   "alpine:latest",
 				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
 				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeOS, types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
+					PkgTypes: []string{
+						types.PkgTypeOS,
+						types.PkgTypeLibrary,
+					},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
@@ -383,7 +608,7 @@ func TestScanner_Scan(t *testing.T) {
 				},
 				Returns: ApplierApplyLayersReturns{
 					Detail: ftypes.ArtifactDetail{
-						OS: &ftypes.OS{
+						OS: ftypes.OS{
 							Family: "alpine",
 							Name:   "3.11",
 						},
@@ -391,14 +616,8 @@ func TestScanner_Scan(t *testing.T) {
 							{
 								Type:     "bundler",
 								FilePath: "/app/Gemfile.lock",
-								Libraries: []ftypes.Package{
-									{
-										Name:    "rails",
-										Version: "4.0.2",
-										Layer: ftypes.Layer{
-											DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
-										},
-									},
+								Packages: []ftypes.Package{
+									railsPkg,
 								},
 							},
 						},
@@ -406,43 +625,48 @@ func TestScanner_Scan(t *testing.T) {
 					Err: analyzer.ErrNoPkgsDetected,
 				},
 			},
-			ospkgDetectExpectations: []OspkgDetectorDetectExpectation{
-				{
-					Args: OspkgDetectorDetectArgs{
-						OsFamily: "alpine",
-						OsName:   "3.11",
-					},
-					Returns: OspkgDetectorDetectReturns{
-						Eosl: false,
-					},
-				},
-			},
-			wantResults: report.Results{
+			wantResults: types.Results{
 				{
 					Target: "alpine:latest (alpine 3.11)",
-					Class:  report.ClassOSPkg,
-					Type:   vulnerability.Alpine,
+					Class:  types.ClassOSPkg,
+					Type:   ftypes.Alpine,
 				},
 				{
 					Target: "/app/Gemfile.lock",
+					Class:  types.ClassLangPkg,
+					Type:   ftypes.Bundler,
+					Packages: ftypes.Packages{
+						railsPkg,
+					},
 					Vulnerabilities: []types.DetectedVulnerability{
 						{
 							VulnerabilityID:  "CVE-2014-0081",
 							PkgName:          "rails",
 							InstalledVersion: "4.0.2",
 							FixedVersion:     "4.0.3, 3.2.17",
+							Status:           dbTypes.StatusFixed,
 							Layer: ftypes.Layer{
 								DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
 							},
+							PrimaryURL: "https://avd.aquasec.com/nvd/cve-2014-0081",
+							Vulnerability: dbTypes.Vulnerability{
+								Title:       "xss",
+								Description: "xss vulnerability",
+								Severity:    "MEDIUM",
+								References: []string{
+									"http://example.com",
+								},
+								LastModifiedDate: lo.ToPtr(time.Date(2020, 2, 1, 1, 1, 0, 0, time.UTC)),
+								PublishedDate:    lo.ToPtr(time.Date(2020, 1, 1, 1, 1, 0, 0, time.UTC)),
+							},
 						},
 					},
-					Class: report.ClassLangPkg,
-					Type:  "bundler",
 				},
 			},
-			wantOS: &ftypes.OS{
+			wantOS: ftypes.OS{
 				Family: "alpine",
 				Name:   "3.11",
+				Eosl:   true,
 			},
 		},
 		{
@@ -451,8 +675,12 @@ func TestScanner_Scan(t *testing.T) {
 				target:   "fedora:27",
 				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
 				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeOS, types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
+					PkgTypes: []string{
+						types.PkgTypeOS,
+						types.PkgTypeLibrary,
+					},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
@@ -462,58 +690,54 @@ func TestScanner_Scan(t *testing.T) {
 				},
 				Returns: ApplierApplyLayersReturns{
 					Detail: ftypes.ArtifactDetail{
-						OS: &ftypes.OS{
+						OS: ftypes.OS{
 							Family: "fedora",
 							Name:   "27",
 						},
 						Applications: []ftypes.Application{
 							{
-								Type:     "bundler",
+								Type:     ftypes.Bundler,
 								FilePath: "/app/Gemfile.lock",
-								Libraries: []ftypes.Package{
-									{
-										Name:    "rails",
-										Version: "4.0.2",
-										Layer: ftypes.Layer{
-											DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
-										},
-									},
+								Packages: []ftypes.Package{
+									railsPkg,
 								},
 							},
 						},
 					},
 				},
 			},
-			ospkgDetectExpectations: []OspkgDetectorDetectExpectation{
+			wantResults: types.Results{
 				{
-					Args: OspkgDetectorDetectArgs{
-						OsFamily: "fedora",
-						OsName:   "27",
-					},
-					Returns: OspkgDetectorDetectReturns{
-						Err: ospkgDetector.ErrUnsupportedOS,
-					},
-				},
-			},
-			wantResults: report.Results{
-				{
-					Target: "/app/Gemfile.lock",
+					Target:   "/app/Gemfile.lock",
+					Class:    types.ClassLangPkg,
+					Type:     ftypes.Bundler,
+					Packages: ftypes.Packages{railsPkg},
 					Vulnerabilities: []types.DetectedVulnerability{
 						{
 							VulnerabilityID:  "CVE-2014-0081",
-							PkgName:          "rails",
-							InstalledVersion: "4.0.2",
+							PkgName:          railsPkg.Name,
+							InstalledVersion: railsPkg.Version,
 							FixedVersion:     "4.0.3, 3.2.17",
+							Status:           dbTypes.StatusFixed,
 							Layer: ftypes.Layer{
-								DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
+								DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+							},
+							PrimaryURL: "https://avd.aquasec.com/nvd/cve-2014-0081",
+							Vulnerability: dbTypes.Vulnerability{
+								Title:       "xss",
+								Description: "xss vulnerability",
+								Severity:    "MEDIUM",
+								References: []string{
+									"http://example.com",
+								},
+								LastModifiedDate: lo.ToPtr(time.Date(2020, 2, 1, 1, 1, 0, 0, time.UTC)),
+								PublishedDate:    lo.ToPtr(time.Date(2020, 1, 1, 1, 1, 0, 0, time.UTC)),
 							},
 						},
 					},
-					Class: report.ClassLangPkg,
-					Type:  ftypes.Bundler,
 				},
 			},
-			wantOS: &ftypes.OS{
+			wantOS: ftypes.OS{
 				Family: "fedora",
 				Name:   "27",
 			},
@@ -524,8 +748,12 @@ func TestScanner_Scan(t *testing.T) {
 				target:   "busybox:latest",
 				layerIDs: []string{"sha256:a6d503001157aedc826853f9b67f26d35966221b158bff03849868ae4a821116"},
 				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeOS, types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
+					PkgTypes: []string{
+						types.PkgTypeOS,
+						types.PkgTypeLibrary,
+					},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
@@ -534,105 +762,98 @@ func TestScanner_Scan(t *testing.T) {
 					BlobIDs: []string{"sha256:a6d503001157aedc826853f9b67f26d35966221b158bff03849868ae4a821116"},
 				},
 				Returns: ApplierApplyLayersReturns{
-					Detail: ftypes.ArtifactDetail{
-						OS: nil,
-					},
 					Err: analyzer.ErrUnknownOS,
 				},
 			},
 			wantResults: nil,
-			wantOS:      nil,
 		},
 		{
-			name: "happy path with only library detection",
+			name: "happy path with only language-specific package detection, excluding direct packages",
 			args: args{
 				target:   "alpine:latest",
-				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
+				layerIDs: []string{"sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33"},
 				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
+					PkgTypes: []string{types.PkgTypeLibrary},
+					PkgRelationships: []ftypes.Relationship{
+						ftypes.RelationshipUnknown,
+						ftypes.RelationshipRoot,
+						ftypes.RelationshipIndirect,
+					},
+					Scanners: types.Scanners{types.VulnerabilityScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
 			applyLayersExpectation: ApplierApplyLayersExpectation{
 				Args: ApplierApplyLayersArgs{
-					BlobIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
+					BlobIDs: []string{"sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33"},
 				},
 				Returns: ApplierApplyLayersReturns{
 					Detail: ftypes.ArtifactDetail{
-						OS: &ftypes.OS{
+						OS: ftypes.OS{
 							Family: "alpine",
 							Name:   "3.11",
 						},
 						Packages: []ftypes.Package{
-							{Name: "musl", Version: "1.2.3"},
+							muslPkg,
 						},
 						Applications: []ftypes.Application{
 							{
 								Type:     "bundler",
 								FilePath: "/app/Gemfile.lock",
-								Libraries: []ftypes.Package{
-									{
-										Name:    "rails",
-										Version: "4.0.2",
-										Layer: ftypes.Layer{
-											DiffID: "sha256:5cb2a5009179b1e78ecfef81a19756328bb266456cf9a9dbbcf9af8b83b735f0",
-										},
-									},
+								Packages: []ftypes.Package{
+									railsPkg,
 								},
 							},
 							{
 								Type:     "composer",
 								FilePath: "/app/composer-lock.json",
-								Libraries: []ftypes.Package{
-									{
-										Name:    "laravel/framework",
-										Version: "6.0.0",
-										Layer: ftypes.Layer{
-											DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
-										},
-									},
+								Packages: []ftypes.Package{
+									laravelPkg, // will be excluded
+									guzzlePkg,
 								},
 							},
 						},
 					},
 				},
 			},
-			wantResults: report.Results{
+			wantResults: types.Results{
 				{
-					Target: "/app/Gemfile.lock",
+					Target:   "/app/Gemfile.lock",
+					Class:    types.ClassLangPkg,
+					Type:     ftypes.Bundler,
+					Packages: ftypes.Packages{railsPkg},
 					Vulnerabilities: []types.DetectedVulnerability{
 						{
 							VulnerabilityID:  "CVE-2014-0081",
-							PkgName:          "rails",
-							InstalledVersion: "4.0.2",
+							PkgName:          railsPkg.Name,
+							InstalledVersion: railsPkg.Version,
 							FixedVersion:     "4.0.3, 3.2.17",
+							Status:           dbTypes.StatusFixed,
 							Layer: ftypes.Layer{
-								DiffID: "sha256:5cb2a5009179b1e78ecfef81a19756328bb266456cf9a9dbbcf9af8b83b735f0",
+								DiffID: "sha256:0ea33a93585cf1917ba522b2304634c3073654062d5282c1346322967790ef33",
+							},
+							PrimaryURL: "https://avd.aquasec.com/nvd/cve-2014-0081",
+							Vulnerability: dbTypes.Vulnerability{
+								Title:       "xss",
+								Description: "xss vulnerability",
+								Severity:    "MEDIUM",
+								References: []string{
+									"http://example.com",
+								},
+								LastModifiedDate: lo.ToPtr(time.Date(2020, 2, 1, 1, 1, 0, 0, time.UTC)),
+								PublishedDate:    lo.ToPtr(time.Date(2020, 1, 1, 1, 1, 0, 0, time.UTC)),
 							},
 						},
 					},
-					Class: report.ClassLangPkg,
-					Type:  ftypes.Bundler,
 				},
 				{
-					Target: "/app/composer-lock.json",
-					Vulnerabilities: []types.DetectedVulnerability{
-						{
-							VulnerabilityID:  "CVE-2021-21263",
-							PkgName:          "laravel/framework",
-							InstalledVersion: "6.0.0",
-							FixedVersion:     "8.22.1, 7.30.3, 6.20.12",
-							Layer: ftypes.Layer{
-								DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
-							},
-						},
-					},
-					Class: report.ClassLangPkg,
-					Type:  ftypes.Composer,
+					Target:   "/app/composer-lock.json",
+					Class:    types.ClassLangPkg,
+					Type:     ftypes.Composer,
+					Packages: ftypes.Packages{guzzlePkg},
 				},
 			},
-			wantOS: &ftypes.OS{
+			wantOS: ftypes.OS{
 				Family: "alpine",
 				Name:   "3.11",
 			},
@@ -643,7 +864,7 @@ func TestScanner_Scan(t *testing.T) {
 				target:   "/app/configs",
 				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
 				options: types.ScanOptions{
-					SecurityChecks: []string{types.SecurityCheckConfig},
+					Scanners: types.Scanners{types.MisconfigScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
@@ -668,17 +889,6 @@ func TestScanner_Scan(t *testing.T) {
 										},
 									},
 								},
-								Exceptions: ftypes.MisconfResults{
-									{
-										Namespace: "main.kubernetes.id100",
-										PolicyMetadata: ftypes.PolicyMetadata{
-											ID:       "ID100",
-											Type:     "Kubernetes Security Check",
-											Title:    "Bad Deployment",
-											Severity: "HIGH",
-										},
-									},
-								},
 								Layer: ftypes.Layer{
 									DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
 								},
@@ -688,7 +898,7 @@ func TestScanner_Scan(t *testing.T) {
 								FilePath: "/app/configs/deployment.yaml",
 								Successes: []ftypes.MisconfResult{
 									{
-										Namespace: "appshield.kubernetes.id200",
+										Namespace: "builtin.kubernetes.id200",
 										PolicyMetadata: ftypes.PolicyMetadata{
 											ID:       "ID200",
 											Type:     "Kubernetes Security Check",
@@ -717,10 +927,10 @@ func TestScanner_Scan(t *testing.T) {
 					},
 				},
 			},
-			wantResults: report.Results{
+			wantResults: types.Results{
 				{
 					Target: "/app/configs/deployment.yaml",
-					Class:  report.ClassConfig,
+					Class:  types.ClassConfig,
 					Type:   ftypes.Kubernetes,
 					Misconfigurations: []types.DetectedMisconfiguration{
 						{
@@ -730,7 +940,7 @@ func TestScanner_Scan(t *testing.T) {
 							Message:   "something bad",
 							Namespace: "main.kubernetes.id100",
 							Severity:  "HIGH",
-							Status:    types.StatusFailure,
+							Status:    types.MisconfStatusFailure,
 							Layer: ftypes.Layer{
 								DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
 							},
@@ -740,13 +950,13 @@ func TestScanner_Scan(t *testing.T) {
 							ID:         "ID200",
 							Title:      "Bad Deployment",
 							Message:    "No issues found",
-							Namespace:  "appshield.kubernetes.id200",
+							Namespace:  "builtin.kubernetes.id200",
 							Severity:   "MEDIUM",
-							PrimaryURL: "https://avd.aquasec.com/appshield/id200",
+							PrimaryURL: "https://avd.aquasec.com/misconfig/id200",
 							References: []string{
-								"https://avd.aquasec.com/appshield/id200",
+								"https://avd.aquasec.com/misconfig/id200",
 							},
-							Status: types.StatusPassed,
+							Status: types.MisconfStatusPassed,
 							Layer: ftypes.Layer{
 								DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
 							},
@@ -755,7 +965,7 @@ func TestScanner_Scan(t *testing.T) {
 				},
 				{
 					Target: "/app/configs/pod.yaml",
-					Class:  report.ClassConfig,
+					Class:  types.ClassConfig,
 					Type:   ftypes.Kubernetes,
 					Misconfigurations: []types.DetectedMisconfiguration{
 						{
@@ -765,19 +975,7 @@ func TestScanner_Scan(t *testing.T) {
 							Message:   "No issues found",
 							Namespace: "main.kubernetes.id300",
 							Severity:  "MEDIUM",
-							Status:    types.StatusFailure,
-							Layer: ftypes.Layer{
-								DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
-							},
-						},
-						{
-							Type:      "Kubernetes Security Check",
-							ID:        "ID100",
-							Title:     "Bad Deployment",
-							Message:   "No issues found",
-							Namespace: "main.kubernetes.id100",
-							Severity:  "HIGH",
-							Status:    types.StatusException,
+							Status:    types.MisconfStatusFailure,
 							Layer: ftypes.Layer{
 								DiffID: "sha256:9922bc15eeefe1637b803ef2106f178152ce19a391f24aec838cbe2e48e73303",
 							},
@@ -792,8 +990,12 @@ func TestScanner_Scan(t *testing.T) {
 				target:   "alpine:latest",
 				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
 				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeOS, types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
+					PkgTypes: []string{
+						types.PkgTypeOS,
+						types.PkgTypeLibrary,
+					},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/happy.yaml"},
@@ -808,68 +1010,14 @@ func TestScanner_Scan(t *testing.T) {
 			wantErr: "failed to apply layers",
 		},
 		{
-			name: "sad path: ospkgDetector.Detect returns an error",
-			args: args{
-				target:   "alpine:latest",
-				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
-				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeOS, types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
-				},
-			},
-			fixtures: []string{"testdata/fixtures/happy.yaml"},
-			applyLayersExpectation: ApplierApplyLayersExpectation{
-				Args: ApplierApplyLayersArgs{
-					BlobIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
-				},
-				Returns: ApplierApplyLayersReturns{
-					Detail: ftypes.ArtifactDetail{
-						OS: &ftypes.OS{
-							Family: "alpine",
-							Name:   "3.11",
-						},
-						Packages: []ftypes.Package{
-							{
-								Name:    "musl",
-								Version: "1.2.3",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-						},
-					},
-				},
-			},
-			ospkgDetectExpectations: []OspkgDetectorDetectExpectation{
-				{
-					Args: OspkgDetectorDetectArgs{
-						OsFamily: "alpine",
-						OsName:   "3.11",
-						Pkgs: []ftypes.Package{
-							{
-								Name:    "musl",
-								Version: "1.2.3",
-								Layer: ftypes.Layer{
-									DiffID: "sha256:ebf12965380b39889c99a9c02e82ba465f887b45975b6e389d42e9e6a3857888",
-								},
-							},
-						},
-					},
-					Returns: OspkgDetectorDetectReturns{
-						Err: errors.New("error"),
-					},
-				},
-			},
-			wantErr: "failed to scan OS packages",
-		},
-		{
 			name: "sad path: library.Detect returns an error",
 			args: args{
 				target:   "alpine:latest",
 				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
 				options: types.ScanOptions{
-					VulnType:       []string{types.VulnTypeLibrary},
-					SecurityChecks: []string{types.SecurityCheckVulnerability},
+					PkgTypes:         []string{types.PkgTypeLibrary},
+					PkgRelationships: ftypes.Relationships,
+					Scanners:         types.Scanners{types.VulnerabilityScanner},
 				},
 			},
 			fixtures: []string{"testdata/fixtures/sad.yaml"},
@@ -879,7 +1027,7 @@ func TestScanner_Scan(t *testing.T) {
 				},
 				Returns: ApplierApplyLayersReturns{
 					Detail: ftypes.ArtifactDetail{
-						OS: &ftypes.OS{
+						OS: ftypes.OS{
 							Family: "alpine",
 							Name:   "3.11",
 						},
@@ -896,7 +1044,7 @@ func TestScanner_Scan(t *testing.T) {
 							{
 								Type:     "bundler",
 								FilePath: "/app/Gemfile.lock",
-								Libraries: []ftypes.Package{
+								Packages: []ftypes.Package{
 									{
 										Name:    "rails",
 										Version: "6.0",
@@ -912,6 +1060,132 @@ func TestScanner_Scan(t *testing.T) {
 			},
 			wantErr: "failed to scan application libraries",
 		},
+		{
+			name: "scan image history",
+			args: args{
+				target:   "alpine:latest",
+				layerIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
+				options: types.ScanOptions{
+					ImageConfigScanners: types.Scanners{types.MisconfigScanner},
+				},
+			},
+			fixtures: []string{"testdata/fixtures/happy.yaml"},
+			applyLayersExpectation: ApplierApplyLayersExpectation{
+				Args: ApplierApplyLayersArgs{
+					BlobIDs: []string{"sha256:5216338b40a7b96416b8b9858974bbe4acc3096ee60acbc4dfb1ee02aecceb10"},
+				},
+				Returns: ApplierApplyLayersReturns{
+					Detail: ftypes.ArtifactDetail{
+						OS: ftypes.OS{
+							Family: ftypes.Alpine,
+							Name:   "3.11",
+						},
+						Misconfigurations: []ftypes.Misconfiguration{
+							{
+								FileType: ftypes.Dockerfile,
+								FilePath: "Dockerfile",
+								Successes: ftypes.MisconfResults{
+									{
+										Namespace: "builtin.dockerfile.DS001",
+										Query:     "data.builtin.dockerfile.DS001.deny",
+										Message:   "",
+										PolicyMetadata: ftypes.PolicyMetadata{
+											ID:                 "DS001",
+											AVDID:              "AVD-DS-0001",
+											Type:               "Dockerfile Security Check",
+											Title:              "':latest' tag used",
+											Description:        "When using a 'FROM' statement you should use a specific tag to avoid uncontrolled behavior when the image is updated.",
+											Severity:           "MEDIUM",
+											RecommendedActions: "Add a tag to the image in the 'FROM' statement",
+										},
+										CauseMetadata: ftypes.CauseMetadata{
+											Provider: "Dockerfile",
+											Service:  "general",
+											Code:     ftypes.Code{},
+										},
+									},
+								},
+								Failures: ftypes.MisconfResults{
+									{
+										Namespace: "builtin.dockerfile.DS002",
+										Query:     "data.builtin.dockerfile.DS002.deny",
+										Message:   "Specify at least 1 USER command in Dockerfile with non-root user as argument",
+										PolicyMetadata: ftypes.PolicyMetadata{
+											ID:                 "DS002",
+											AVDID:              "AVD-DS-0002",
+											Type:               "Dockerfile Security Check",
+											Title:              "Image user should not be 'root'",
+											Description:        "Running containers with 'root' user can lead to a container escape situation. It is a best practice to run containers as non-root users, which can be done by adding a 'USER' statement to the Dockerfile.",
+											Severity:           "HIGH",
+											RecommendedActions: "Add 'USER <non root user name>' line to the Dockerfile",
+										},
+										CauseMetadata: ftypes.CauseMetadata{
+											Provider: "Dockerfile",
+											Service:  "general",
+											Code:     ftypes.Code{},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			wantResults: types.Results{
+				{
+					Target: "Dockerfile",
+					Class:  types.ClassConfig,
+					Type:   ftypes.Dockerfile,
+					Misconfigurations: []types.DetectedMisconfiguration{
+						{
+							Namespace:   "builtin.dockerfile.DS002",
+							Query:       "data.builtin.dockerfile.DS002.deny",
+							Message:     "Specify at least 1 USER command in Dockerfile with non-root user as argument",
+							Type:        "Dockerfile Security Check",
+							ID:          "DS002",
+							AVDID:       "AVD-DS-0002",
+							Title:       "Image user should not be 'root'",
+							Description: "Running containers with 'root' user can lead to a container escape situation. It is a best practice to run containers as non-root users, which can be done by adding a 'USER' statement to the Dockerfile.",
+							Severity:    "HIGH",
+							Resolution:  "Add 'USER <non root user name>' line to the Dockerfile",
+							Status:      types.MisconfStatusFailure,
+							PrimaryURL:  "https://avd.aquasec.com/misconfig/ds002",
+							References:  []string{"https://avd.aquasec.com/misconfig/ds002"},
+							CauseMetadata: ftypes.CauseMetadata{
+								Provider: "Dockerfile",
+								Service:  "general",
+								Code:     ftypes.Code{},
+							},
+						},
+						{
+							Namespace:   "builtin.dockerfile.DS001",
+							Query:       "data.builtin.dockerfile.DS001.deny",
+							Message:     "No issues found",
+							Type:        "Dockerfile Security Check",
+							ID:          "DS001",
+							AVDID:       "AVD-DS-0001",
+							Title:       "':latest' tag used",
+							Description: "When using a 'FROM' statement you should use a specific tag to avoid uncontrolled behavior when the image is updated.",
+							Severity:    "MEDIUM",
+							Resolution:  "Add a tag to the image in the 'FROM' statement",
+							Status:      types.MisconfStatusPassed,
+							CauseMetadata: ftypes.CauseMetadata{
+								Provider: "Dockerfile",
+								Service:  "general",
+								Code:     ftypes.Code{},
+							},
+							PrimaryURL: "https://avd.aquasec.com/misconfig/ds001",
+							References: []string{"https://avd.aquasec.com/misconfig/ds001"},
+						},
+					},
+				},
+			},
+			wantOS: ftypes.OS{
+				Family: "alpine",
+				Name:   "3.11",
+				Eosl:   false,
+			},
+		},
 	}
 
 	for _, tt := range tests {
@@ -922,24 +1196,18 @@ func TestScanner_Scan(t *testing.T) {
 			applier := new(MockApplier)
 			applier.ApplyApplyLayersExpectation(tt.applyLayersExpectation)
 
-			ospkgDetector := new(MockOspkgDetector)
-			ospkgDetector.ApplyDetectExpectations(tt.ospkgDetectExpectations)
-
-			s := NewScanner(applier, ospkgDetector)
-			gotResults, gotOS, err := s.Scan(tt.args.target, "", tt.args.layerIDs, tt.args.options)
+			s := NewScanner(applier, ospkg.NewScanner(), langpkg.NewScanner(), vulnerability.NewClient(db.Config{}))
+			gotResults, gotOS, err := s.Scan(context.Background(), tt.args.target, "", tt.args.layerIDs, tt.args.options)
 			if tt.wantErr != "" {
-				require.NotNil(t, err, tt.name)
-				require.Contains(t, err.Error(), tt.wantErr, tt.name)
+				require.ErrorContains(t, err, tt.wantErr, tt.name)
 				return
-			} else {
-				require.NoError(t, err, tt.name)
 			}
 
+			require.NoError(t, err, tt.name)
 			assert.Equal(t, tt.wantResults, gotResults)
 			assert.Equal(t, tt.wantOS, gotOS)
 
 			applier.AssertExpectations(t)
-			ospkgDetector.AssertExpectations(t)
 		})
 	}
 }
