@@ -3,10 +3,11 @@
 package scanner
 
 import (
-	fanaltypes "github.com/aquasecurity/fanal/types"
+	"context"
+
 	mock "github.com/stretchr/testify/mock"
 
-	report "github.com/aquasecurity/trivy/pkg/report"
+	fanaltypes "github.com/aquasecurity/trivy/pkg/fanal/types"
 
 	types "github.com/aquasecurity/trivy/pkg/types"
 )
@@ -17,6 +18,8 @@ type MockDriver struct {
 }
 
 type DriverScanArgs struct {
+	Ctx              context.Context
+	CtxAnything      bool
 	Target           string
 	TargetAnything   bool
 	ImageID          string
@@ -28,8 +31,8 @@ type DriverScanArgs struct {
 }
 
 type DriverScanReturns struct {
-	Results report.Results
-	OsFound *fanaltypes.OS
+	Results types.Results
+	OsFound fanaltypes.OS
 	Err     error
 }
 
@@ -40,6 +43,11 @@ type DriverScanExpectation struct {
 
 func (_m *MockDriver) ApplyScanExpectation(e DriverScanExpectation) {
 	var args []interface{}
+	if e.Args.CtxAnything {
+		args = append(args, mock.Anything)
+	} else {
+		args = append(args, e.Args.Ctx)
+	}
 	if e.Args.TargetAnything {
 		args = append(args, mock.Anything)
 	} else {
@@ -69,31 +77,31 @@ func (_m *MockDriver) ApplyScanExpectations(expectations []DriverScanExpectation
 	}
 }
 
-// Scan provides a mock function with given fields: target, imageID, layerIDs, options
-func (_m *MockDriver) Scan(target string, artifactKey string, blobKeys []string, options types.ScanOptions) (report.Results, *fanaltypes.OS, error) {
-	ret := _m.Called(target, artifactKey, blobKeys, options)
+// Scan provides a mock function with given fields: ctx, target, imageID, layerIDs, options
+func (_m *MockDriver) Scan(ctx context.Context, target string, artifactKey string, blobKeys []string, options types.ScanOptions) (types.Results, fanaltypes.OS, error) {
+	ret := _m.Called(ctx, target, artifactKey, blobKeys, options)
 
-	var r0 report.Results
-	if rf, ok := ret.Get(0).(func(string, string, []string, types.ScanOptions) report.Results); ok {
-		r0 = rf(target, artifactKey, blobKeys, options)
+	var r0 types.Results
+	if rf, ok := ret.Get(0).(func(context.Context, string, string, []string, types.ScanOptions) types.Results); ok {
+		r0 = rf(ctx, target, artifactKey, blobKeys, options)
 	} else {
 		if ret.Get(0) != nil {
-			r0 = ret.Get(0).(report.Results)
+			r0 = ret.Get(0).(types.Results)
 		}
 	}
 
-	var r1 *fanaltypes.OS
-	if rf, ok := ret.Get(1).(func(string, string, []string, types.ScanOptions) *fanaltypes.OS); ok {
-		r1 = rf(target, artifactKey, blobKeys, options)
+	var r1 fanaltypes.OS
+	if rf, ok := ret.Get(1).(func(context.Context, string, string, []string, types.ScanOptions) fanaltypes.OS); ok {
+		r1 = rf(ctx, target, artifactKey, blobKeys, options)
 	} else {
 		if ret.Get(1) != nil {
-			r1 = ret.Get(1).(*fanaltypes.OS)
+			r1 = ret.Get(1).(fanaltypes.OS)
 		}
 	}
 
 	var r2 error
-	if rf, ok := ret.Get(2).(func(string, string, []string, types.ScanOptions) error); ok {
-		r2 = rf(target, artifactKey, blobKeys, options)
+	if rf, ok := ret.Get(2).(func(context.Context, string, string, []string, types.ScanOptions) error); ok {
+		r2 = rf(ctx, target, artifactKey, blobKeys, options)
 	} else {
 		r2 = ret.Error(2)
 	}
