@@ -6,12 +6,16 @@ import (
 	"github.com/google/wire"
 	"golang.org/x/xerrors"
 
-	"github.com/aquasecurity/fanal/artifact"
-	aimage "github.com/aquasecurity/fanal/artifact/image"
-	flocal "github.com/aquasecurity/fanal/artifact/local"
-	"github.com/aquasecurity/fanal/artifact/remote"
-	"github.com/aquasecurity/fanal/image"
-	ftypes "github.com/aquasecurity/fanal/types"
+	"github.com/aquasecurity/trivy/pkg/cache"
+	"github.com/aquasecurity/trivy/pkg/clock"
+	"github.com/aquasecurity/trivy/pkg/fanal/artifact"
+	aimage "github.com/aquasecurity/trivy/pkg/fanal/artifact/image"
+	flocal "github.com/aquasecurity/trivy/pkg/fanal/artifact/local"
+	"github.com/aquasecurity/trivy/pkg/fanal/artifact/repo"
+	"github.com/aquasecurity/trivy/pkg/fanal/artifact/sbom"
+	"github.com/aquasecurity/trivy/pkg/fanal/artifact/vm"
+	"github.com/aquasecurity/trivy/pkg/fanal/image"
+	ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
 	"github.com/aquasecurity/trivy/pkg/log"
 	"github.com/aquasecurity/trivy/pkg/report"
 	"github.com/aquasecurity/trivy/pkg/rpc/client"
@@ -19,8 +23,17 @@ import (
 	"github.com/aquasecurity/trivy/pkg/types"
 )
 
+///////////////
+// Standalone
+///////////////
+
 // StandaloneSuperSet is used in the standalone mode
 var StandaloneSuperSet = wire.NewSet(
+	// Cache
+	cache.New,
+	wire.Bind(new(cache.ArtifactCache), new(cache.Cache)),
+	wire.Bind(new(cache.LocalArtifactCache), new(cache.Cache)),
+
 	local.SuperSet,
 	wire.Bind(new(Driver), new(local.Scanner)),
 	NewScanner,
@@ -28,8 +41,7 @@ var StandaloneSuperSet = wire.NewSet(
 
 // StandaloneDockerSet binds docker dependencies
 var StandaloneDockerSet = wire.NewSet(
-	types.GetDockerOption,
-	image.NewDockerImage,
+	image.NewContainerImage,
 	aimage.NewArtifact,
 	StandaloneSuperSet,
 )
@@ -43,33 +55,78 @@ var StandaloneArchiveSet = wire.NewSet(
 
 // StandaloneFilesystemSet binds filesystem dependencies
 var StandaloneFilesystemSet = wire.NewSet(
-	flocal.NewArtifact,
+	flocal.ArtifactSet,
 	StandaloneSuperSet,
 )
 
 // StandaloneRepositorySet binds repository dependencies
 var StandaloneRepositorySet = wire.NewSet(
-	remote.NewArtifact,
+	repo.ArtifactSet,
 	StandaloneSuperSet,
 )
 
+// StandaloneSBOMSet binds sbom dependencies
+var StandaloneSBOMSet = wire.NewSet(
+	sbom.NewArtifact,
+	StandaloneSuperSet,
+)
+
+// StandaloneVMSet binds vm dependencies
+var StandaloneVMSet = wire.NewSet(
+	vm.ArtifactSet,
+	StandaloneSuperSet,
+)
+
+/////////////////
+// Client/Server
+/////////////////
+
 // RemoteSuperSet is used in the client mode
 var RemoteSuperSet = wire.NewSet(
-	aimage.NewArtifact,
-	client.SuperSet,
+	// Cache
+	cache.NewRemoteCache,
+	wire.Bind(new(cache.ArtifactCache), new(*cache.RemoteCache)), // No need for LocalArtifactCache
+
+	client.NewScanner,
+	wire.Value([]client.Option(nil)),
 	wire.Bind(new(Driver), new(client.Scanner)),
 	NewScanner,
 )
 
+// RemoteFilesystemSet binds filesystem dependencies for client/server mode
+var RemoteFilesystemSet = wire.NewSet(
+	flocal.ArtifactSet,
+	RemoteSuperSet,
+)
+
+// RemoteRepositorySet binds repository dependencies for client/server mode
+var RemoteRepositorySet = wire.NewSet(
+	repo.ArtifactSet,
+	RemoteSuperSet,
+)
+
+// RemoteSBOMSet binds sbom dependencies for client/server mode
+var RemoteSBOMSet = wire.NewSet(
+	sbom.NewArtifact,
+	RemoteSuperSet,
+)
+
+// RemoteVMSet binds vm dependencies for client/server mode
+var RemoteVMSet = wire.NewSet(
+	vm.ArtifactSet,
+	RemoteSuperSet,
+)
+
 // RemoteDockerSet binds remote docker dependencies
 var RemoteDockerSet = wire.NewSet(
-	types.GetDockerOption,
-	image.NewDockerImage,
+	aimage.NewArtifact,
+	image.NewContainerImage,
 	RemoteSuperSet,
 )
 
 // RemoteArchiveSet binds remote archive dependencies
 var RemoteArchiveSet = wire.NewSet(
+	aimage.NewArtifact,
 	image.NewArchiveImage,
 	RemoteSuperSet,
 )
@@ -82,43 +139,54 @@ type Scanner struct {
 
 // Driver defines operations of scanner
 type Driver interface {
-	Scan(target string, artifactKey string, blobKeys []string, options types.ScanOptions) (
-		results report.Results, osFound *ftypes.OS, err error)
+	Scan(ctx context.Context, target, artifactKey string, blobKeys []string, options types.ScanOptions) (
+		results types.Results, osFound ftypes.OS, err error)
 }
 
 // NewScanner is the factory method of Scanner
 func NewScanner(driver Driver, ar artifact.Artifact) Scanner {
-	return Scanner{driver: driver, artifact: ar}
+	return Scanner{
+		driver:   driver,
+		artifact: ar,
+	}
 }
 
 // ScanArtifact scans the artifacts and returns results
-func (s Scanner) ScanArtifact(ctx context.Context, options types.ScanOptions) (report.Report, error) {
+func (s Scanner) ScanArtifact(ctx context.Context, options types.ScanOptions) (types.Report, error) {
 	artifactInfo, err := s.artifact.Inspect(ctx)
 	if err != nil {
-		return report.Report{}, xerrors.Errorf("failed analysis: %w", err)
+		return types.Report{}, xerrors.Errorf("failed analysis: %w", err)
 	}
+	defer func() {
+		if err := s.artifact.Clean(artifactInfo); err != nil {
+			log.Warn("Failed to clean the artifact",
+				log.String("artifact", artifactInfo.Name), log.Err(err))
+		}
+	}()
 
-	results, osFound, err := s.driver.Scan(artifactInfo.Name, artifactInfo.ID, artifactInfo.BlobIDs, options)
+	results, osFound, err := s.driver.Scan(ctx, artifactInfo.Name, artifactInfo.ID, artifactInfo.BlobIDs, options)
 	if err != nil {
-		return report.Report{}, xerrors.Errorf("scan failed: %w", err)
+		return types.Report{}, xerrors.Errorf("scan failed: %w", err)
 	}
 
-	if osFound != nil && osFound.Eosl {
-		log.Logger.Warnf("This OS version is no longer supported by the distribution: %s %s", osFound.Family, osFound.Name)
-		log.Logger.Warnf("The vulnerability detection may be insufficient because security updates are not provided")
+	ptros := &osFound
+	if osFound.Detected() && osFound.Eosl {
+		log.Warn("This OS version is no longer supported by the distribution",
+			log.String("family", string(osFound.Family)), log.String("version", osFound.Name))
+		log.Warn("The vulnerability detection may be insufficient because security updates are not provided")
+	} else if !osFound.Detected() {
+		ptros = nil
 	}
 
-	// Layer makes sense only when scanning container images
-	if artifactInfo.Type != ftypes.ArtifactContainerImage {
-		removeLayer(results)
-	}
-
-	return report.Report{
+	return types.Report{
 		SchemaVersion: report.SchemaVersion,
+		CreatedAt:     clock.Now(ctx),
 		ArtifactName:  artifactInfo.Name,
 		ArtifactType:  artifactInfo.Type,
-		Metadata: report.Metadata{
-			OS:          osFound,
+		Metadata: types.Metadata{
+			OS: ptros,
+
+			// Container image
 			ImageID:     artifactInfo.ImageMetadata.ID,
 			DiffIDs:     artifactInfo.ImageMetadata.DiffIDs,
 			RepoTags:    artifactInfo.ImageMetadata.RepoTags,
@@ -126,21 +194,6 @@ func (s Scanner) ScanArtifact(ctx context.Context, options types.ScanOptions) (r
 			ImageConfig: artifactInfo.ImageMetadata.ConfigFile,
 		},
 		Results: results,
+		BOM:     artifactInfo.BOM,
 	}, nil
-}
-
-func removeLayer(results report.Results) {
-	for i := range results {
-		result := results[i]
-
-		for j := range result.Packages {
-			result.Packages[j].Layer = ftypes.Layer{}
-		}
-		for j := range result.Vulnerabilities {
-			result.Vulnerabilities[j].Layer = ftypes.Layer{}
-		}
-		for j := range result.Misconfigurations {
-			result.Misconfigurations[j].Layer = ftypes.Layer{}
-		}
-	}
 }
