@@ -1,23 +1,24 @@
 package types
 
-import ftypes "github.com/aquasecurity/fanal/types"
+import ftypes "github.com/aquasecurity/trivy/pkg/fanal/types"
 
 // DetectedMisconfiguration holds detected misconfigurations
 type DetectedMisconfiguration struct {
-	Type        string             `json:",omitempty"`
-	ID          string             `json:",omitempty"`
-	Title       string             `json:",omitempty"`
-	Description string             `json:",omitempty"`
-	Message     string             `json:",omitempty"`
-	Namespace   string             `json:",omitempty"`
-	Query       string             `json:",omitempty"`
-	Resolution  string             `json:",omitempty"`
-	Severity    string             `json:",omitempty"`
-	PrimaryURL  string             `json:",omitempty"`
-	References  []string           `json:",omitempty"`
-	Status      MisconfStatus      `json:",omitempty"`
-	Layer       ftypes.Layer       `json:",omitempty"`
-	IacMetadata ftypes.IacMetadata `json:",omitempty"`
+	Type          string               `json:",omitempty"`
+	ID            string               `json:",omitempty"`
+	AVDID         string               `json:",omitempty"`
+	Title         string               `json:",omitempty"`
+	Description   string               `json:",omitempty"`
+	Message       string               `json:",omitempty"`
+	Namespace     string               `json:",omitempty"`
+	Query         string               `json:",omitempty"`
+	Resolution    string               `json:",omitempty"`
+	Severity      string               `json:",omitempty"`
+	PrimaryURL    string               `json:",omitempty"`
+	References    []string             `json:",omitempty"`
+	Status        MisconfStatus        `json:",omitempty"`
+	Layer         ftypes.Layer         `json:",omitempty"`
+	CauseMetadata ftypes.CauseMetadata `json:",omitempty"`
 
 	// For debugging
 	Traces []string `json:",omitempty"`
@@ -27,12 +28,14 @@ type DetectedMisconfiguration struct {
 type MisconfStatus string
 
 const (
-	// StatusPassed represents successful status
-	StatusPassed MisconfStatus = "PASS"
+	// MisconfStatusPassed represents successful status
+	MisconfStatusPassed MisconfStatus = "PASS"
 
-	// StatusFailure represents failure status
-	StatusFailure MisconfStatus = "FAIL"
+	// MisconfStatusFailure represents failure status
+	MisconfStatusFailure MisconfStatus = "FAIL"
 
-	// StatusException Passed represents the status of exception
-	StatusException MisconfStatus = "EXCEPTION"
+	// MisconfStatusException Passed represents the status of exception
+	MisconfStatusException MisconfStatus = "EXCEPTION"
 )
+
+func (DetectedMisconfiguration) findingType() FindingType { return FindingTypeMisconfiguration }
