@@ -645,9 +645,12 @@ func sanitizeBaseURL(baseURL string) string {
 
 // baseServicePath composes the path prefix for the service (without <Method>).
 // e.g.: baseServicePath("/twirp", "my.pkg", "MyService")
-//       returns => "/twirp/my.pkg.MyService/"
+//
+//	returns => "/twirp/my.pkg.MyService/"
+//
 // e.g.: baseServicePath("", "", "MyService")
-//       returns => "/MyService/"
+//
+//	returns => "/MyService/"
 func baseServicePath(prefix, pkg, service string) string {
 	fullServiceName := service
 	if pkg != "" {
@@ -1091,36 +1094,48 @@ func callClientError(ctx context.Context, h *twirp.ClientHooks, err twirp.Error)
 }
 
 var twirpFileDescriptor0 = []byte{
-	// 486 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x84, 0x53, 0x3d, 0x6f, 0xdb, 0x30,
-	0x10, 0x85, 0x64, 0xc7, 0x1f, 0xa7, 0xa2, 0x4e, 0x88, 0xb6, 0x50, 0x12, 0xb4, 0x35, 0x3c, 0xb4,
-	0x46, 0x07, 0x09, 0x56, 0x80, 0x76, 0xe8, 0xd4, 0x8f, 0x0c, 0x19, 0x8a, 0x04, 0x74, 0xd0, 0xa1,
-	0x8b, 0x40, 0x51, 0x8c, 0x43, 0x84, 0x16, 0x15, 0x92, 0x32, 0xa0, 0xa9, 0xff, 0x23, 0xbf, 0xb6,
-	0x10, 0x29, 0x05, 0xb5, 0x03, 0x23, 0x93, 0x74, 0xef, 0x1e, 0xef, 0xde, 0xbd, 0x23, 0xe1, 0x58,
-	0x95, 0x34, 0xd6, 0x94, 0x14, 0x05, 0x53, 0xb1, 0x66, 0x6a, 0xc3, 0x29, 0x8b, 0x4a, 0x25, 0x8d,
-	0x44, 0x87, 0x46, 0xf1, 0x4d, 0x1d, 0xb5, 0xc9, 0x68, 0xb3, 0x38, 0xf9, 0xbc, 0xe2, 0xe6, 0xb6,
-	0xca, 0x22, 0x2a, 0xd7, 0x31, 0xb9, 0xaf, 0x88, 0x66, 0xb4, 0x52, 0xdc, 0xd4, 0xb1, 0x65, 0xc6,
-	0x4d, 0x29, 0x2a, 0xd7, 0x6b, 0x59, 0x6c, 0x57, 0x9a, 0x3d, 0x78, 0x10, 0x2c, 0x29, 0x29, 0x30,
-	0xbb, 0xaf, 0x98, 0x36, 0xe8, 0x0d, 0x0c, 0x0c, 0x51, 0x2b, 0x66, 0x42, 0x6f, 0xea, 0xcd, 0xc7,
-	0xb8, 0x8d, 0xd0, 0x7b, 0x08, 0x88, 0x32, 0xfc, 0x86, 0x50, 0x93, 0xf2, 0x3c, 0xf4, 0x6d, 0x12,
-	0x3a, 0xe8, 0x22, 0x47, 0xc7, 0x30, 0xca, 0x84, 0xcc, 0x52, 0x9e, 0xeb, 0xb0, 0x37, 0xed, 0xcd,
-	0xc7, 0x78, 0xd8, 0xc4, 0x17, 0xb9, 0x46, 0x5f, 0x60, 0x28, 0x4b, 0xc3, 0x65, 0xa1, 0xc3, 0xfe,
-	0xd4, 0x9b, 0x07, 0xc9, 0xdb, 0x68, 0x57, 0x7f, 0xd4, 0x68, 0xb8, 0x74, 0x24, 0xdc, 0xb1, 0x67,
-	0x7f, 0x9d, 0xb6, 0x16, 0x47, 0xa7, 0x30, 0xde, 0x54, 0xa2, 0x48, 0x4d, 0x5d, 0xb2, 0xd0, 0xb3,
-	0x3d, 0x46, 0x0d, 0x70, 0x5d, 0x97, 0x0c, 0x7d, 0x84, 0x49, 0x37, 0x73, 0x4a, 0x6f, 0x19, 0xbd,
-	0xd3, 0xa1, 0x6f, 0x29, 0x2f, 0x3b, 0xf8, 0x87, 0x45, 0xd1, 0x27, 0x38, 0x12, 0x5c, 0x9b, 0x94,
-	0x08, 0x91, 0x96, 0x84, 0xde, 0x91, 0x15, 0x6b, 0x14, 0x7b, 0xf3, 0x11, 0x9e, 0x34, 0x89, 0x6f,
-	0x42, 0x5c, 0xb5, 0xf0, 0x2c, 0x87, 0x17, 0xce, 0x1c, 0x5d, 0xca, 0x42, 0x33, 0x34, 0x05, 0x5f,
-	0x6a, 0xeb, 0x4c, 0x90, 0x1c, 0xb6, 0x43, 0x38, 0x5b, 0xa3, 0xcb, 0x25, 0xf6, 0xa5, 0x46, 0x09,
-	0x0c, 0x15, 0xd3, 0x95, 0x30, 0xce, 0x85, 0x20, 0x09, 0x9f, 0xce, 0x8a, 0x2d, 0x01, 0x77, 0xc4,
-	0xd9, 0x83, 0x0f, 0x03, 0x87, 0xed, 0xb5, 0xff, 0x1c, 0x26, 0xcd, 0xa4, 0x4c, 0x91, 0x8c, 0x0b,
-	0x6e, 0x38, 0x73, 0xd3, 0x05, 0xc9, 0xe9, 0xb6, 0x8a, 0xdf, 0xff, 0x91, 0x6a, 0xbc, 0x7b, 0x06,
-	0x5d, 0xc3, 0xd1, 0x9a, 0x6b, 0x2a, 0x8b, 0x1b, 0xbe, 0xaa, 0x14, 0xe9, 0x76, 0xd2, 0x14, 0xfa,
-	0xb0, 0x5d, 0xe8, 0x27, 0x33, 0x8c, 0x1a, 0x96, 0xff, 0xda, 0xa1, 0xe3, 0xa7, 0x05, 0xd0, 0x2b,
-	0x38, 0xa0, 0x82, 0x68, 0x1d, 0x0e, 0xac, 0x66, 0x17, 0x20, 0x04, 0x7d, 0xbb, 0xa8, 0x9e, 0x05,
-	0xed, 0x3f, 0x5a, 0xc0, 0xe8, 0xd1, 0xf2, 0x03, 0xdb, 0xf6, 0xf5, 0x76, 0xdb, 0xd6, 0x79, 0xfc,
-	0x48, 0x4b, 0xae, 0x60, 0xb8, 0x74, 0xd6, 0xa1, 0x73, 0xe8, 0x37, 0xbf, 0x68, 0xcf, 0xf5, 0x69,
-	0xaf, 0xf0, 0xc9, 0xbb, 0x7d, 0x69, 0xb7, 0xc4, 0xef, 0x67, 0x7f, 0x16, 0xcf, 0x3f, 0x96, 0xf6,
-	0xf0, 0xd7, 0xf6, 0x9b, 0x0d, 0xec, 0x73, 0x39, 0xfb, 0x17, 0x00, 0x00, 0xff, 0xff, 0xf1, 0x14,
-	0x4b, 0x5a, 0x95, 0x03, 0x00, 0x00,
+	// 685 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x7c, 0x54, 0x51, 0x6f, 0x1a, 0x47,
+	0x10, 0x16, 0x1c, 0x86, 0x63, 0xa8, 0x6a, 0xbc, 0x6a, 0xad, 0x35, 0xae, 0x5b, 0xc4, 0x43, 0x85,
+	0x54, 0x09, 0x6a, 0xdc, 0xaa, 0x4d, 0xf2, 0x16, 0xdb, 0x89, 0x1c, 0x25, 0xb2, 0xb5, 0x58, 0x79,
+	0xc8, 0x0b, 0x59, 0xf6, 0xc6, 0xe7, 0x15, 0xc7, 0xed, 0x79, 0x77, 0x0f, 0x89, 0xff, 0x91, 0xa7,
+	0xfc, 0xaf, 0xfc, 0x9f, 0x68, 0xf7, 0x0e, 0x64, 0x30, 0xce, 0x13, 0x3b, 0x33, 0xdf, 0xcc, 0x7c,
+	0xcc, 0x7c, 0x37, 0x70, 0xa4, 0x33, 0x31, 0x34, 0x82, 0xa7, 0x29, 0xea, 0xa1, 0x41, 0xbd, 0x90,
+	0x02, 0x07, 0x99, 0x56, 0x56, 0x91, 0xb6, 0xd5, 0x72, 0xb1, 0x1c, 0x94, 0xc1, 0xc1, 0xe2, 0xb4,
+	0x43, 0x1d, 0x58, 0xa8, 0xf9, 0x5c, 0xa5, 0x9b, 0xd8, 0xde, 0xd7, 0x0a, 0xb4, 0xc6, 0x82, 0xa7,
+	0x0c, 0x1f, 0x72, 0x34, 0x96, 0x1c, 0x42, 0xdd, 0x72, 0x1d, 0xa3, 0xa5, 0x95, 0x6e, 0xa5, 0xdf,
+	0x64, 0xa5, 0x45, 0xfe, 0x80, 0x16, 0xd7, 0x56, 0xde, 0x71, 0x61, 0x27, 0x32, 0xa2, 0x55, 0x1f,
+	0x84, 0x95, 0xeb, 0x2a, 0x22, 0x47, 0x10, 0x4e, 0x13, 0x35, 0x9d, 0xc8, 0xc8, 0xd0, 0xa0, 0x1b,
+	0xf4, 0x9b, 0xac, 0xe1, 0xec, 0xab, 0xc8, 0x90, 0xff, 0xa0, 0xa1, 0x32, 0x2b, 0x55, 0x6a, 0x68,
+	0xad, 0x5b, 0xe9, 0xb7, 0x46, 0x27, 0x83, 0x6d, 0x86, 0x03, 0xc7, 0xe1, 0xba, 0x00, 0xb1, 0x15,
+	0xba, 0xd7, 0x85, 0xf0, 0xbd, 0x14, 0x98, 0x1a, 0x34, 0xe4, 0x17, 0xd8, 0x4b, 0xf9, 0x1c, 0x0d,
+	0xad, 0xf8, 0xe2, 0x85, 0xd1, 0xfb, 0x12, 0x14, 0xf4, 0xcb, 0x54, 0x72, 0x0c, 0xcd, 0x6c, 0x16,
+	0x4f, 0xec, 0x32, 0x5b, 0x23, 0xc3, 0x6c, 0x16, 0xdf, 0x3a, 0x9b, 0x74, 0x20, 0x2c, 0x3b, 0x1a,
+	0x5a, 0x2d, 0x62, 0x2b, 0x9b, 0x08, 0x20, 0x49, 0xd1, 0x6a, 0x22, 0xb8, 0xc5, 0x58, 0x69, 0x89,
+	0x8e, 0x6e, 0xd0, 0x6f, 0x8d, 0xfe, 0xf9, 0x21, 0xdd, 0x41, 0x49, 0xf1, 0x7c, 0x9d, 0x76, 0x99,
+	0x5a, 0xbd, 0x64, 0x07, 0xc9, 0xb6, 0x9f, 0xf4, 0xa1, 0x2d, 0x53, 0x91, 0xe4, 0x11, 0x4e, 0x22,
+	0x5c, 0x4c, 0x22, 0xcc, 0x0c, 0xdd, 0xeb, 0x56, 0xfa, 0x21, 0xfb, 0xb9, 0xf4, 0x5f, 0xe0, 0xe2,
+	0x02, 0x33, 0x43, 0xfe, 0x82, 0x03, 0xf7, 0x3f, 0x34, 0x26, 0xdc, 0x37, 0xb9, 0x97, 0x99, 0xa1,
+	0x75, 0xcf, 0xb9, 0x9d, 0xcd, 0x62, 0xf6, 0xd8, 0x4f, 0xfa, 0x50, 0x8f, 0xa4, 0xb1, 0x5a, 0xd1,
+	0x86, 0x1f, 0x6f, 0xbb, 0xe4, 0x5b, 0x2c, 0x7c, 0x70, 0x3d, 0x66, 0x65, 0xbc, 0xf3, 0x19, 0x0e,
+	0x77, 0xb3, 0x25, 0x6d, 0x08, 0x66, 0xb8, 0x2c, 0x97, 0xee, 0x9e, 0xe4, 0x6f, 0xd8, 0x5b, 0xf0,
+	0x24, 0x47, 0xbf, 0xeb, 0xd6, 0xa8, 0xf3, 0x74, 0x08, 0xab, 0xdd, 0xb0, 0x02, 0xf8, 0xb2, 0xfa,
+	0x7f, 0xe5, 0x5d, 0x2d, 0x0c, 0xda, 0xb5, 0x5e, 0x04, 0x3f, 0x15, 0xa2, 0x32, 0x99, 0x4a, 0x0d,
+	0x92, 0x2e, 0x54, 0x95, 0xf1, 0xc5, 0x77, 0xb1, 0xab, 0x2a, 0x43, 0x46, 0xd0, 0xd0, 0x68, 0xf2,
+	0xc4, 0x16, 0xea, 0x69, 0x8d, 0xe8, 0xd3, 0x7e, 0xcc, 0x03, 0xd8, 0x0a, 0xd8, 0xfb, 0x16, 0x40,
+	0xbd, 0xf0, 0x3d, 0x2b, 0xdb, 0x4b, 0xd8, 0x5f, 0xe4, 0x49, 0x8a, 0x9a, 0x4f, 0x65, 0x22, 0xad,
+	0xdb, 0x69, 0xd5, 0x97, 0x3f, 0xde, 0x64, 0xf1, 0xf1, 0x11, 0x68, 0xc9, 0xb6, 0x73, 0xc8, 0x2d,
+	0x1c, 0xcc, 0xa5, 0x11, 0x2a, 0xbd, 0x93, 0x71, 0xae, 0xf9, 0x4a, 0xcb, 0xae, 0xd0, 0x9f, 0x9b,
+	0x85, 0x2e, 0xd0, 0xa2, 0xb0, 0x18, 0x7d, 0xd8, 0x82, 0xb3, 0xa7, 0x05, 0x9c, 0xa4, 0x45, 0xc2,
+	0x8d, 0x5b, 0xac, 0xe3, 0x5c, 0x18, 0x84, 0x40, 0xcd, 0xc9, 0x97, 0x06, 0xde, 0xe9, 0xdf, 0xe4,
+	0x14, 0xc2, 0x8c, 0x8b, 0x19, 0x8f, 0xd1, 0x09, 0xc6, 0xb5, 0xfd, 0x75, 0xb3, 0xed, 0x4d, 0x11,
+	0x65, 0x6b, 0x18, 0x79, 0x0b, 0x6d, 0x91, 0x1b, 0xab, 0xe6, 0x13, 0x8d, 0x46, 0xe5, 0x5a, 0xa0,
+	0xa1, 0x0d, 0x9f, 0xfa, 0xdb, 0x66, 0xea, 0xb9, 0x47, 0xb1, 0x12, 0xc4, 0xf6, 0xc5, 0x86, 0x6d,
+	0xc8, 0xbf, 0xd0, 0x30, 0x28, 0x34, 0x5a, 0x43, 0xc3, 0x5d, 0xa3, 0x1b, 0xfb, 0xe0, 0x1b, 0x99,
+	0x46, 0x32, 0x8d, 0xd9, 0x0a, 0x4b, 0x5e, 0x40, 0x58, 0x7e, 0x00, 0x86, 0x36, 0x7d, 0xde, 0xc9,
+	0xee, 0x49, 0x95, 0x2a, 0x62, 0x6b, 0xf8, 0xe8, 0x06, 0x1a, 0xe3, 0x62, 0xeb, 0xe4, 0x12, 0x6a,
+	0xee, 0x49, 0x9e, 0xb9, 0x18, 0xe5, 0xd5, 0xea, 0xfc, 0xfe, 0x5c, 0xb8, 0xd0, 0xdf, 0xeb, 0xb3,
+	0x4f, 0xa7, 0xb1, 0xb4, 0xf7, 0xf9, 0xd4, 0x35, 0x1f, 0xf2, 0x87, 0x9c, 0x1b, 0x14, 0xb9, 0x96,
+	0x76, 0x39, 0xf4, 0x89, 0xc3, 0x47, 0xc7, 0xf4, 0x55, 0xf9, 0x3b, 0xad, 0xfb, 0x0b, 0x79, 0xf6,
+	0x3d, 0x00, 0x00, 0xff, 0xff, 0x40, 0xd0, 0xe7, 0xda, 0x6a, 0x05, 0x00, 0x00,
 }
